% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_fix.R
\name{adapt_list_length}
\alias{adapt_list_length}
\title{Adapt schema max list length based on hint}
\usage{
adapt_list_length(view, hint)
}
\description{
Adapt schema max list length based on hint
}
\keyword{internal}
