% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_fix.R
\name{adapt_view}
\alias{adapt_view}
\title{Adapt view}
\usage{
adapt_view(view, max_tries = 5L)
}
\arguments{
\item{view}{Synapse view id.}

\item{max_tries}{Number of tries. Most views should only have accumulated 1-2 issues, so a default of 5 should be reasonable.}
}
\description{
When a view schema and data are mismatched, the view cannot be built and therefore cannot be queried.
The most common causes (likely accounting for 98\%+ of instances combined) will be the max size/length issues.
This will iteratively update the schema based exactly on whatever the server is saying to do* for the
sizing and list length issues, until the view is functional again and the query works.
However, if the issue is not one of these, this will fail because handlers for other, rarer problems are currently not implemented.
}
\details{
*Note: We apply iterative fixes because that's just how server works currently in surfacing repair recommendations.
}
