% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{add_activity}
\alias{add_activity}
\title{Add activity to entity}
\usage{
add_activity(entity, act_name, act_executed, used_inputs)
}
\arguments{
\item{entity}{Synapse entity id.}

\item{act_name}{Name of activity.}

\item{act_executed}{Reference to the the activity executed (URL preferred).}

\item{used_inputs}{Vector of inputs for this act, e.g. syn ids, links to other data sources, etc.}
}
\description{
Util for adding activity info to a file entity.
See also https://help.synapse.org/docs/Provenance.1972470373.html
}
