% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{add_activity_batch}
\alias{add_activity_batch}
\title{Add activity to multiple entities}
\usage{
add_activity_batch(entities, act_name, act_executed, used_inputs)
}
\arguments{
\item{entities}{Vector or list of entities.}

\item{act_name}{Vector or list of activity name.}

\item{act_executed}{Vector or list of reference activity executed.}

\item{used_inputs}{Vector or list of inputs for each entity.}
}
\description{
Wrapper provenance function that does a little more work to
expand many-to-many mappings to create records of entity, activity, and input.
}
