% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R
\name{add_default_folders}
\alias{add_default_folders}
\title{Create default folders}
\usage{
add_default_folders(
  project,
  folders = c("Analysis", "Milestone Reports", "Raw Data")
)
}
\arguments{
\item{project}{The project Synapse id or object.}

\item{folders}{Names of the standard set of folders.}
}
\description{
A convenience wrapper around \code{make_folder} with NF defaults.
}
