% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_study.R
\name{add_new_study_meta}
\alias{add_new_study_meta}
\title{Helpers ---------------------------------------------------------------------#
Add meta for new studies as annotations}
\usage{
add_new_study_meta(id, study_meta)
}
\arguments{
\item{id}{Id of container representing the study, usually a Synapse project.}

\item{study_meta}{Study meta as a list.}
}
\description{
Put \emph{selected} metadata into Synapse annotations for the study project entity.
}
\keyword{internal}
