% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_people_table.R
\name{add_people_from_table}
\alias{add_people_from_table}
\title{Update the People table from a source Table or View column}
\usage{
add_people_from_table(
  people_table_id,
  people_column,
  source_table_id,
  source_column,
  dry_run = T
)
}
\arguments{
\item{dry_run}{Default = TRUE. Skips upload to table and instead prints study tibble.}

\item{fileview_id}{The synapse id of a fileview. Must have the desired annotations in the schema, and must have the files to annotate included in the scope. Must have write access to the files you want to re-annotate.}

\item{annotation_key}{A character string of the annotation you'd like to switch from a delimited string to a stringlist.}

\item{sep}{Default = ",". The delimiter in the character string.}

\item{trimws}{Default = TRUE. Remove white space at the beginning and end of list items (e.g. "NF1, NF2" and "NF1,NF2" will yield the same STRING_LIST result).}
}
\value{
If dry_run == T, returns study tibble and skips upload.
}
\description{
Update the People table from a source Table or View column
}
