% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_utils.R
\name{add_prov_rna_seq}
\alias{add_prov_rna_seq}
\title{Add provenance for nextflow RNA-seq results}
\usage{
add_prov_rna_seq(
  samplesheet,
  syn_out,
  workflow = "STAR and Salmon",
  workflow_link = "https://nf-co.re/rnaseq/3.4/output#star-and-salmon",
  dry_run = TRUE
)
}
\arguments{
\item{samplesheet}{A local file or syn id of samplesheet.}

\item{syn_out}{Syn id of syn output destination (folder) with files of interest.}

\item{workflow}{Workflow name (activity name).}

\item{workflow_link}{Workflow link.}

\item{dry_run}{Whether to set provenance or only return data.}
}
\description{
Composes together some utils to do provenance annotation in one step.
}
