% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_publication_from_pubmed.R
\name{add_publication_from_pubmed}
\alias{add_publication_from_pubmed}
\title{Add a publication to the publication table}
\usage{
add_publication_from_pubmed(
  pmid,
  study_id,
  disease_focus = c(""),
  manifestation = c(""),
  publication_table_id,
  study_table_id,
  dry_run = T
)
}
\arguments{
\item{pmid}{PubMed ID (\emph{not} PMCID) of the publication to be added.}

\item{study_id}{Synapse id(s) of the study that are associated with the publication.}

\item{disease_focus}{(Optional) The disease focus(s) that are associated with the publication.}

\item{manifestation}{(Optional) The manifestation(s) that are associated with the publication.}

\item{publication_table_id}{Synapse id of the portal publication table. Must have write access.}

\item{study_table_id}{Synapse id of the portal study table. Need read access.}

\item{dry_run}{Default = TRUE. Skips upload to table and instead prints formatted publication metadata.}
}
\value{
If dry_run == T, returns publication metadata to be added.
}
\description{
Requires that the publication be in PubMed to auto-derive metadata such as authors, title, etc.
In contrast, \code{disease_focus} and \code{manifestation} need to be supplemented by the curator.
The \code{study_id} is used to get consistent \code{studyName} and \code{fundingAgency} from study table without manual input.
}
\examples{
\dontrun{
add_publication_from_pubmed(
               pmid = "33574490",
               study_id = "syn2343195",
               disease_focus = c("Neurofibromatosis"),
               manifestation = c("Meningioma"),
               publication_table_id = "syn16857542",
               study_table_id = "syn16787123")
}
}
