% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_publication_from_pubmed.R
\name{add_publications_from_file}
\alias{add_publications_from_file}
\title{Add a batch of publications from spreadsheet}
\usage{
add_publications_from_file(
  file,
  publication_table_id,
  study_table_id,
  list_sep = "|",
  dry_run = TRUE
)
}
\arguments{
\item{file}{Spreadsheet (.csv/.tsv) with pubs to add should have \code{pmid}, \code{studyId}, \code{diseaseFocus}, \code{manifestation}.
\code{pmid} is one per row and unique, rest can be \code{list_sep} vals.}

\item{publication_table_id}{Synapse id of the portal publication table. Must have write access.}

\item{study_table_id}{Synapse id of the portal study table. Need read access.}

\item{list_sep}{Delimiter character used to separate list columns.}

\item{dry_run}{Default = TRUE. Skips upload to table and instead prints formatted publication metadata.}
}
\description{
Add a batch of publications from spreadsheet
}
