% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pubmed_publication.R
\name{add_pubmed_publication}
\alias{add_pubmed_publication}
\title{Add a publication to the publication table.}
\usage{
add_pubmed_publication(
  publication_table_id,
  pmid = NA,
  doi = NA,
  study_name,
  study_id,
  funding_agency,
  disease_focus,
  manifestation,
  dry_run = T
)
}
\arguments{
\item{publication_table_id}{The synapse id of the portal publication table. Must have write access.}

\item{pmid}{The PubMed ID (\emph{not} PMCID) of the publication to be added. Provide this \emph{or} a DOI.}

\item{doi}{The DOI (\emph{not} PMCID) of the publication to be added. Provide this \emph{or} a PMID. Even if providing a DOI, the publication must be indexed in PubMed.}

\item{study_name}{The name(s) of the study that are associated with the publication.}

\item{study_id}{The synapse id(s) of the study that are associated with the publication.}

\item{funding_agency}{The funding agency(s) that are associated with the publication.}

\item{disease_focus}{The disease focus(s) that are associated with the publication.}

\item{manifestation}{The manifestation(s) that are associated with the publication.}

\item{dry_run}{Default = TRUE. Skips upload to table and instead prints formatted publication metadata.}
}
\value{
If dry_run == T, returns publication metadata to be added.
}
\description{
Add a publication to the publication table. Requires that the publication be listed in PubMed.
}
\examples{
add_pubmed_publication(publication_table_id = 'syn16857542',
              doi = '10.1074/jbc.RA120.014960',
               study_name = c(toJSON("Synodos NF2")),
               study_id = c(toJSON("syn2343195")),
               funding_agency = c(toJSON("CTF")),
               disease_focus = c(toJSON("Neurofibromatosis 2")),
               manifestation = c(toJSON("Meningioma")),
               dry_run = F)
}
