% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_study.R
\name{add_study_summary}
\alias{add_study_summary}
\title{Add studyId-summary key-value only}
\usage{
add_study_summary(study_id, summary, table_id = "syn16787123")
}
\description{
Add studyId-summary key-value only
}
\keyword{internal}
