% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_fix.R
\name{adjust_list_length}
\alias{adjust_list_length}
\title{Adjust schema max list length based on hint}
\usage{
adjust_list_length(view, hint, check_byte_budget = TRUE)
}
\description{
Adjust schema max list length based on hint
}
\keyword{internal}
