% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_fix.R
\name{adjust_string_size}
\alias{adjust_string_size}
\title{Adjust schema max size based on hint}
\usage{
adjust_string_size(view, hint, check_byte_budget = TRUE)
}
\description{
Note: For STRING cols, the hard limit is 1000 char size,
though at 250 using LARGETEXT is officially recommended, so possibly at that breakpoint
this should just create the different column type instead of increasing size.
}
\keyword{internal}
