% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_annotation.R
\name{annotate_auxiliary}
\alias{annotate_auxiliary}
\title{Recursive annotation of nextflow outputs}
\usage{
annotate_auxiliary(output, workflow_ref, assay = "whole exome sequencing")
}
\arguments{
\item{output}{One or more nextflow output folders, must be a named vector.
This can be a single top-level \code{publishDir} to automate annotations for all outputs, or a couple of folders for more selective annotation.}

\item{workflow_ref}{The base workflow reference used for \code{workflowLink}. If given top-level folder, this can be extracted automatically, otherwise must be passed in.}

\item{assay}{Assay related to data outputs.}
}
\description{
Recursive annotation of nextflow outputs
}
