% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_annotation_utils.R
\name{annotate_called_variants}
\alias{annotate_called_variants}
\title{Annotate somatic or germline variants output}
\usage{
annotate_called_variants(metadata, workflow_link, verbose = TRUE)
}
\arguments{
\item{metadata}{Metadata table to build upon.}

\item{workflow_link}{Workflow link to most specific part of workflow generating these data.}

\item{verbose}{Give verbose reports for what's happening.}
}
\description{
Given a manifest, annotate data as variant data (\code{vcf} or \code{maf}).
}
\details{
\code{maf} files use the same template but with different default values.
If in the future \code{maf}s require a significantly different template,
then this should be factored out into a separate annotation function.
}
