% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_utils.R
\name{annotate_cnv}
\alias{annotate_cnv}
\title{Annotate CNV outputs from CNVKit}
\usage{
annotate_cnv(
  output_dataset,
  input_dataset,
  workflow = "CNVkit",
  workflowLink = "https://nf-co.re/sarek/3.1/docs/output#cnvkit",
  dry_run = TRUE
)
}
\arguments{
\item{output_dataset}{Id of dataset, which should be created first via \code{nf_cnv_dataset}.}

\item{input_dataset}{Input dataset used in the workflow.}

\item{workflow}{Optional name for workflow, defaults to CNVkit.}

\item{workflowLink}{Optional corresponding link to workflow, defaults to known used version.}

\item{dry_run}{Whether to return a version submit manifest.}
}
\description{
Annotate CNV outputs from CNVKit
}
