% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_utils.R
\name{annotate_expression}
\alias{annotate_expression}
\title{Annotate processed expression output}
\usage{
annotate_expression(
  sample_io,
  template = "bts:ProcessedExpressionTemplate",
  schema =
    "https://raw.githubusercontent.com/nf-osi/nf-metadata-dictionary/main/NF.jsonld",
  verbose = TRUE,
  dry_run = TRUE
)
}
\arguments{
\item{sample_io}{Table mapping input to outputs, where outputs are expected to be .sf files.}

\item{template}{(Optional) URI of template in data model to use, prefixed if needed.
Can specify different model/version, but in some cases may not work well.}

\item{schema}{Path (URL or local) to file from which schema will be read, or schema as list object.}

\item{verbose}{Give verbose reports for what's happening.}

\item{dry_run}{Whether to apply annotations.}
}
\description{
Annotate processed expression output
}
