% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_annotation_utils.R
\name{annotate_processed}
\alias{annotate_processed}
\title{Annotate processed data}
\usage{
annotate_processed(metadata, ...)
}
\arguments{
\item{metadata}{Metadata.}

\item{...}{Other parameters to pass to the called annotation function.}
}
\description{
Annotate as some processed data type inferred from output.
}
