% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_annotation_utils.R
\name{annotate_quantified_expression}
\alias{annotate_quantified_expression}
\title{Annotate quantified expression output}
\usage{
annotate_quantified_expression(metadata, workflow_link, verbose = TRUE)
}
\arguments{
\item{metadata}{Metadata table to build upon.}

\item{workflow_link}{Workflow link to most specific part of workflow generating these data.}

\item{verbose}{Give verbose reports for what's happening.}
}
\description{
Given a manifest, annotate data as level 3 processed expression data,
using defaults from star_salmon processing.
Returns a "partial" manifest, which can be adjusted as needed,
e.g. add additional comments or batch info.
}
