% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_utils_reports.R
\name{annotate_reports_sarek}
\alias{annotate_reports_sarek}
\title{Annotate Sarek reports}
\usage{
annotate_reports_sarek(syn_out, project, dry_run)
}
\arguments{
\item{syn_out}{The \code{Reports} output folder which is set as scope for fileview.}

\item{project}{Project in which to put fileview.}

\item{dry_run}{Whether to submit annotations or just return the manifest.}
}
\description{
First runs \code{map_reports_sarek} under the hood.
}
