% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotate_with_manifest}
\alias{annotate_with_manifest}
\title{Set annotations from a manifest}
\usage{
annotate_with_manifest(manifest, ignore_na = TRUE, ignore_blank = TRUE)
}
\arguments{
\item{manifest}{A table manifest. Needs to contain \code{entityId}.}

\item{ignore_na}{Whether to ignore annotations that are \code{NA}; default TRUE.}

\item{ignore_blank}{Whether to ignore annotations that are that empty strings; default TRUE.}
}
\description{
The \href{https://help.synapse.org/docs/Managing-Custom-Metadata-at-Scale.2004254976.html}{Synapse docs}
suggest doing batch annotations through a fileview. However, it is often simpler to
modify or set new annotations directly given a table of just the entities (rows) and props (cols) we want.
This is similar to how schematic works, except without any validation
(so it works best for power-users who know the data model very well).
Some desired defaults are taken into account,
such as not submitting key-values with \code{NA} and empty strings.
}
