% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_annotation_utils.R
\name{annotate_with_samtools_stats}
\alias{annotate_with_samtools_stats}
\title{Make annotations from samtools stats}
\usage{
annotate_with_samtools_stats(meta, samtools_stats_file = NULL, verbose = TRUE)
}
\arguments{
\item{meta}{Data to which tool stats will be added as additional meta.}

\item{samtools_stats_file}{Path to file/syn id of file with samtools stats produced by the workflow.}

\item{verbose}{Whether to output detailed debugging messages.}
}
\description{
Extracts a subset of \href{http://www.htslib.org/doc/samtools-stats.html}{samtools stats}
Regarding the selection of stats, see the Genomic Data Commons (GDC) model for
\href{https://docs.gdc.cancer.gov/Data_Dictionary/viewer/#?view=table-definition-view&id=aligned_reads}{AlignedReads}
}
