% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_annotation_utils.R
\name{annotation_rule}
\alias{annotation_rule}
\title{Match to output-specific annotation function}
\usage{
annotation_rule(outputFrom, which = c("format_as", "annotate_as", "template"))
}
\description{
This encodes logic for annotation and checks, e.g.
DeepVariant is Germline variant calling only, Mutect2 is for Somatic variant calling only,
while FreeBayes and Strelka2 can be applied to both; see
https://raw.githubusercontent.com/nf-core/sarek/3.4.2//docs/images/sarek_workflow.png
}
\keyword{internal}
