% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{append_kv}
\alias{append_kv}
\title{Append key-value pair dependent on value being given}
\usage{
append_kv(x, key, value, quote = FALSE)
}
\arguments{
\item{quote}{Quote the value (needed if contains conflict YAML characters).}
}
\description{
Append key-value pair dependent on value being given
}
\keyword{internal}
