% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{as_clinical_file_type}
\alias{as_clinical_file_type}
\title{Export specific type of clinical data}
\usage{
as_clinical_file_type(
  df,
  clinical_type = c("SAMPLE", "PATIENT"),
  m,
  publish_dir = ".",
  verbose = TRUE
)
}
\arguments{
\item{df}{A \code{data.frame} representing clinical dataset to publicize.}

\item{clinical_type}{\code{SAMPLE} or \code{PATIENT}}

\item{m}{A reference mapping object. See \code{use_ref_map}.}

\item{publish_dir}{Directory path to write to, defaults to current.}

\item{verbose}{Report where file has been written.}
}
\description{
Export specific type of clinical data
}
\keyword{internal}
