% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{as_coll_items}
\alias{as_coll_items}
\title{As collection items}
\usage{
as_coll_items(ids, item_version = NULL)
}
\arguments{
\item{ids}{Ids of entities to make into dataset items.}

\item{item_version}{Integer for version that will be used for all items, e.g. 1.
If NULL, this will look up the latest version for each id and use that.}
}
\description{
Helper taking entity ids to create records used for dataset items or dataset collection items.
Collection items have the form \code{list(entityId = id, versionNumber = x)}.
}
\keyword{internal}
