% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrams.R
\name{as_mmd_link}
\alias{as_mmd_link}
\title{Generate notation for mermaid.js link}
\usage{
as_mmd_link(n1, n2, directional = TRUE, style = "solid")
}
\arguments{
\item{n1}{Id for node at one end.}

\item{n2}{Id for node at other end.}

\item{directional}{Boolean option for diretional arrow from n1 to n2.}

\item{style}{Option for "solid" or "dash"-style links.}
}
\description{
Generate notation for mermaid.js link
}
\keyword{internal}
