% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal_charts.R
\name{as_mmd_node}
\alias{as_mmd_node}
\title{Helper function for rendering nodes}
\usage{
as_mmd_node(entity, named = FALSE, class = c("Project", "Dataset", "Folder"))
}
\description{
Helper function for rendering nodes
}
\keyword{internal}
