% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_reference.R
\name{asExpressionUnit}
\alias{asExpressionUnit}
\title{Match expression unit}
\usage{
asExpressionUnit(filename)
}
\arguments{
\item{filename}{Filename.}
}
\description{
Match expression unit
}
\keyword{internal}
