% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_study_data_types.R
\name{assign_study_data_types}
\alias{assign_study_data_types}
\title{Summarize data types for the study}
\usage{
assign_study_data_types(
  study_table_id,
  fileview_id,
  id_col = "studyId",
  attribute = "dataType",
  dry_run = TRUE
)
}
\arguments{
\item{study_table_id}{Synapse ID of the portal study table/view that lists relevant studies in column \code{id} or \code{studyId}.}

\item{fileview_id}{Synapse ID of the portal fileview.}

\item{id_col}{Name of the id column in \code{study_table_id} and \code{fileview_id}.}

\item{attribute}{The attribute that we are rolling up; name should not contain spaces.
Defaults to \code{dataType}. Must be queryable in \code{fileview_id}.}

\item{dry_run}{Default = TRUE. Skips updating the annotation and instead returns annotation object(s) list.}
}
\description{
Data types are summarized, or "rolled-up", for the study based on its child file annotations.
This summary is added as (and will overwrite the current) \code{dataType} annotation for the study.
Contrast this with \code{update_study_annotations}, where study-level annotations are rolled down to child files.
}
\examples{
\dontrun{
assign_study_data_types(study_table_id = 'syn52694652',
                        fileview_id = 'syn16858331',
                        id_col = 'studyId',
                        id_col = 'studyId',
                        dry_run = T)
}
}
