% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_wiki_links.R
\name{bad_url}
\alias{bad_url}
\title{Helper function to check urls}
\usage{
bad_url(url)
}
\arguments{
\item{url}{A character vector of one or more URLs.}
}
\value{
Result vector of same size with values "bad" or "ok".
}
\description{
Check whether URL(s) return a HTTP error, indicating a broken link.
Note that this uses \code{curl} under the hood, which may give timeout errors and therefore false positives for some links that are valid but take too long to resolve.
}
