% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_utils.R
\name{bare_syn_id}
\alias{bare_syn_id}
\title{Extract synapse id from URI or other string}
\usage{
bare_syn_id(uri)
}
\arguments{
\item{uri}{URI or string containing embedded Synapse id.}
}
\description{
Extract synapse id from URI or other string
}
\keyword{internal}
