% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{bind_schema}
\alias{bind_schema}
\title{Wrapper for JSON schema binding}
\usage{
bind_schema(id, schema_id, derived_annotations = FALSE)
}
\arguments{
\item{id}{Id of entity to which schema will be bound}

\item{schema_id}{Schema id as registered on Synapse.
Examples for "latest" or explicitly versioned schemas: "org.synapse.nf-portalstudy", "org.synapse.nf-rnaseqtemplate-10.2.0".}

\item{derived_annotations}{Whether to enabled derived annotations.
Default \code{FALSE} as this is the API default.}
}
\description{
Binds a JSON schema to a Synapse entity. If a schema is already bound to the entity,
running this function again will replace the existing schema binding with the new one.
}
\details{
See https://help.synapse.org/docs/JSON-Schemas.3107291536.html
}
\examples{
\dontrun{
# Bind a dataset schema to a folder
bind_schema(id = "syn12345678", schema_id = "org.synapse.nf-rnaseqtemplate-10.2.0")

# Bind with derived annotations enabled
bind_schema(id = "syn12345678",
            schema_id = "org.synapse.nf-protocol",
            derived_annotations = TRUE)

# Replace an existing schema binding with a new one
bind_schema(id = "syn12345678", schema_id = "org.synapse.nf-rnaseqtemplate-11.0.0")
}
}
