% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrams.R
\name{bipartite_mmd_template}
\alias{bipartite_mmd_template}
\title{Simple bipartite representation in mermaid charts}
\usage{
bipartite_mmd_template(
  nodeset1,
  nodeset2,
  nodeset1_title = "INPUT",
  nodeset2_title = "OUTPUT",
  links = ""
)
}
\arguments{
\item{nodeset1}{Character vector of one or more node ids. If named, nodes will use names instead of ids as labels.}

\item{nodeset2}{Character vector of one or more node ids. If named, nodes will use names instead of ids as labels.}

\item{nodeset1_title}{Title for nodeset1.}

\item{nodeset2_title}{Title for nodeset2.}

\item{links}{Optional, character vector of edges between nodes.}
}
\description{
Simple bipartite representation in mermaid charts
}
