% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_utils.R
\name{build_annotation_derivatives}
\alias{build_annotation_derivatives}
\title{Build annotations for derived data}
\usage{
build_annotation_derivatives(
  sample_io,
  template,
  schema,
  format,
  verbose = TRUE
)
}
\description{
Intended to be used in specific context by more specialized utils
to get annotations on derived (processed) data.
Can't update assets from this directly because result is considered insufficient.
}
\details{
See \code{\link{inherit_input_annotations}} for
usage in less restrictive context.
}
\keyword{internal}
