% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_mod.R
\name{button_widget}
\alias{button_widget}
\title{Generate button widget for a Synapse wiki}
\usage{
button_widget(label, url, align = c("None", "Left", "Right", "Center"))
}
\arguments{
\item{label}{Button label text.}

\item{url}{URL that the button will link to.}

\item{align}{Button alignment, can be one of "None", "Left", "Right", or "Center" (defaults to "None").}
}
\description{
Generate markup for a button widget for a Synapse project wiki.
Refer to widget docs at \url{https://help.synapse.org/docs/Wikis.1975746682.html#Wikis-WikiWidgets}.
Buttons should be created sparingly and strategically.
See \code{\link{remove_button}} in case of future regret.
}
