% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_related_studies.R
\name{calc_study_dist_dtm}
\alias{calc_study_dist_dtm}
\title{Calculate study distance based on summary text}
\usage{
calc_study_dist_dtm(studies)
}
\arguments{
\item{studies}{A \code{data.frame} where each row is a "document"; should have \code{summary} and \code{studyId}.}
}
\description{
There are different measures of similarity;
this gives cosine similarity based on summary text, which is then converted to a distance matrix.
In the future, other methods may be used for comparison or in ensemble.
}
\keyword{internal}
