% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_related_studies.R
\name{calculate_related_studies}
\alias{calculate_related_studies}
\title{Calculate and add related studies to study table.}
\usage{
calculate_related_studies(study_table_id, n_clust, dry_run = TRUE)
}
\arguments{
\item{study_table_id}{The synapse id of the portal study table. Must have write access.}

\item{n_clust}{The target number of clusters to generate. In practice, I've found that the number of total summaries divided by 3 is a good starting point (100 studies = 33 clusters).}

\item{dry_run}{Default = TRUE. Skips upload to table and instead prints study tibble.}
}
\value{
If dry_run == T, returns study tibble and skips upload.
}
\description{
Processes study summary text to identify clusters of related studies. Calculates tf-idf values for 1 and 2 length ngrams, and clusters studies using the ward.D clustering method. Updates data in a STRINGLIST column "relatedStudies."
}
\examples{
\dontrun{ 
 calculate_related_studies(study_table_id = "syn16787123",
                           n_clust = 36,
                           dry_run = T)
}
}
