% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbioportal_study.R
\name{cbp_add_cna}
\alias{cbp_add_cna}
\title{Export and add CNA (seg) data to cBioPortal dataset}
\usage{
cbp_add_cna(cna_data)
}
\arguments{
\item{cna_data}{Synapse id of CNA data file, currently only handles \code{.seg} file.}
}
\description{
This should be run in an existing dataset package root.
}
