% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbioportal.R
\name{cbp_add_maf}
\alias{cbp_add_maf}
\title{Export and add mutations data to cBioPortal dataset}
\usage{
cbp_add_maf(maf_data, verbose = TRUE)
}
\arguments{
\item{maf_data}{Synapse id of \verb{merged maf} file for public release.}

\item{verbose}{Whether to provide informative messages throughout.}
}
\description{
This should be run in an existing dataset package root.
}
\details{
Get merged maf file that represents filtered subset of \code{maf}s containing only (non-germline) data OK to release publicly.
This needs to be packaged with other files like this
\href{https://github.com/cBioPortal/datahub/tree/1e03ea6ab5e0ddd497ecf349cbee7d50aeebcd5e/public/msk_ch_2020}{example of a public mutations dataset}.
}
