% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbioportal_study.R
\name{cbp_datatypes}
\alias{cbp_datatypes}
\title{Enumerate combinations of valid cBP data types and data subtypes and helper utils if available}
\usage{
cbp_datatypes()
}
\description{
https://docs.cbioportal.org/file-formats/
}
\keyword{internal}
