% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbioportal.R
\name{cbp_new_cancer_type}
\alias{cbp_new_cancer_type}
\title{Create reference file for new cancer type}
\usage{
cbp_new_cancer_type(type_of_cancer, name, color, parent_type_of_cancer)
}
\arguments{
\item{type_of_cancer}{Id for new cancer type, e.g. "cnf".}

\item{name}{Full name for new cancer type, e.g. "Cutaneous Neurofibroma"}

\item{color}{Color name for new cancer; https://en.wikipedia.org/wiki/Web_colors#X11_color_names.}

\item{parent_type_of_cancer}{Id of existing parent, e.g. "nfib" for Neurofibroma.}
}
\description{
Helper for creating reference for new cancer subtype which does not already exist.
https://docs.cbioportal.org/file-formats/#cancer-type
}
