% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbioportal_study.R
\name{cbp_new_study}
\alias{cbp_new_study}
\title{Initialize a new cBioPortal study dataset}
\usage{
cbp_new_study(
  cancer_study_identifier,
  name,
  type_of_cancer,
  description =
    "The data are contributed by researchers funded by the Neurofibromatosis Therapeutic Acceleration Program (NTAP). The reprocessing of the raw data is managed by the NF Open Science Initiative (https://nf.synapse.org/).",
  short_name = NULL,
  citation = NULL,
  pmid = NULL,
  groups = "PUBLIC",
  add_global_case_list = TRUE,
  validate = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{cancer_study_identifier}{Cancer study identifier in format such as \code{nst_nfosi_ntap_2022}.}

\item{name}{Name of the study, e.g. "Malignant Peripheral Nerve Sheath Tumor (NF-OSI, 2022)".}

\item{type_of_cancer}{Id for type of cancer. If \code{validate} is TRUE, this is one of the things validated with warning if mismatched.}

\item{description}{Description of the study, defaults to a generic description that can be edited later.}

\item{short_name}{(Optional) Short name for the study.}

\item{citation}{(Optional) A relevant citation, e.g. "TCGA, Nature 2012".}

\item{pmid}{(Optional) One or more relevant pubmed ids (comma-separated, no whitespace); if used, citation cannot be \code{NULL}.}

\item{groups}{(Optional) Defaults to "PUBLIC" for use with public cBioPortal;
otherwise, use group names that makes sense with the configuration of your cBioPortal instance.}

\item{add_global_case_list}{(Optional) Use \code{NULL} to ignore, default is \code{TRUE} for an "All samples" case list( to be generated automatically.}

\item{validate}{Validate against public cBioPortal configuration. Default \code{TRUE},
but might want to set to \code{FALSE} especially if using a custom cBioPortal instance with different configuration.}

\item{verbose}{Verbosity level.}
}
\description{
Create a new directory with a basic required \href{https://docs.cbioportal.org/file-formats/#meta-file}{study meta file},
much like how we'd create a new R package and put a DESCRIPTION file in it.
}
