% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_fix.R
\name{check_byte_budget_col_swap}
\alias{check_byte_budget_col_swap}
\title{Check byte budget when swapping cols in schema}
\usage{
check_byte_budget_col_swap(schema, ref_col, new_col)
}
\description{
Check byte budget when swapping cols in schema
}
\keyword{internal}
