% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_testing_utils.R
\name{check_libtype_validity}
\alias{check_libtype_validity}
\title{Check output strandedness matches samplesheet strandedness}
\usage{
check_libtype_validity(multiqc_id)
}
\arguments{
\item{multiqc_id}{The Synapse ID of the "multiqc_data.json" output file}
}
\description{
This function allows you to (post-workflow) check that the library type inferred by
during processing matches the "strandedness" provided for each sample in the workflow sample sheet.
}
\examples{
\dontrun{
 out <- check_libtype_validity(multiqc_id = 'syn29530566') #should pass
 out <- check_libtype_validity(multiqc_id = 'syn34573377') #should fail
}
}
