% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncBP.R
\name{check_maf_release}
\alias{check_maf_release}
\title{Check maf file for release}
\usage{
check_maf_release(merged_maf, samplesheet)
}
\arguments{
\item{merged_maf}{Maf file as a \code{data.table}.}

\item{samplesheet}{Samplesheet as a \code{data.table}.}
}
\value{
Returns \code{NULL} if everything OK, else the sample ids that don't match expectations.
}
\description{
Currently, this is a simple check to make sure released samples are expected.
It may be extended later on as needed.
}
