% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_testing_utils.R
\name{check_readpair_validity}
\alias{check_readpair_validity}
\title{Check fastq read pair matches samplesheet read pair assignment.}
\usage{
check_readpair_validity(
  samplesheet,
  parse_fun = function(x) gsub("_T[0-9]$", "", x)
)
}
\arguments{
\item{samplesheet}{A local file or syn id of samplesheet.}

\item{parse_fun}{Function implementing how to parse samples in samplesheet.}
}
\description{
Read pairs are often encoded in the name of the file. Here, we check that if encoded in the name of the file,
that the samplesheet read pair (e.g. _1 or _2) matches
}
\examples{
\dontrun{
 check_readpair_validity('syn39542932')
 check_readpair_validity('syn29530880')
}
}
