% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_testing_utils.R
\name{checked_message}
\alias{checked_message}
\title{Format checked message notification}
\usage{
checked_message(string)
}
\arguments{
\item{string}{Character string with check prepended.}
}
\description{
For nicely displaying that some step was successfully completed.
}
\keyword{internal}
