% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{cite_dataset}
\alias{cite_dataset}
\title{Generate dataset citation}
\usage{
cite_dataset(dataset_id, format = "Scientific Data", output = c("markdown"))
}
\arguments{
\item{dataset_id}{Dataset id.}

\item{format}{Currently just "Scientific Data" format.}

\item{output}{Currently just markdown format. There are many ways to
generate LaTeX or HTML from markdown.}
}
\description{
This is currently more for demo purposes, to check how well current metadata
could be formatted into citation text. Datasets within the official
Portal Collection should work well enough, while there are no guarantees for
unofficial/community-contributed datasets.
}
\keyword{internal}
