% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_utils.R
\name{copy}
\alias{copy}
\title{Create copy of entity}
\usage{
copy(
  entity,
  destination_id,
  skip_copy_wiki_page = FALSE,
  skip_copy_annotations = FALSE
)
}
\arguments{
\item{entity}{Entity to copy.}

\item{destination_id}{Id of destination project/container that entity will be copied to.}

\item{skip_copy_wiki_page}{Whether to skip copying wiki; defaults FALSE.}

\item{skip_copy_annotations}{Whether to skip copying annotations; defaults FALSE.}
}
\description{
Create a copy of syn entity; mostly used to create a copy on which to test out changes.
See https://python-docs.synapse.org/build/html/synapseutils.html?highlight=copy#synapseutils.copy_functions.copy
}
