% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_utils.R
\name{copy_table}
\alias{copy_table}
\title{Copy a table}
\usage{
copy_table(table_id, destination_id)
}
\arguments{
\item{table_id}{Id of table to copy.}

\item{destination_id}{Parent project id for the copy.}
}
\description{
Copy a table. One of the most common use cases is testing, to avoid modifying a "production" table.
}
