% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_mod.R
\name{data_curator_app_subpage}
\alias{data_curator_app_subpage}
\title{Create NF Data Curator App subpage}
\usage{
data_curator_app_subpage(project_id, dry_run = TRUE)
}
\arguments{
\item{project_id}{ID of the owner Synapse project.}

\item{dry_run}{Whether to return a wiki object only without actually performing update.}
}
\description{
Convenience method to create a subpage with the default buttons for the annotation app and docs.
This is a highly specific method and expected to have a limited lifespan.
}
