% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{delete_provenance}
\alias{delete_provenance}
\title{Remove provenance info}
\usage{
delete_provenance(entities)
}
\arguments{
\item{entities}{Vector or list of entities.}
}
\description{
Remove provenance info
}
