% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_publication_from_pubmed.R
\name{.add_publication_from_pubmed}
\alias{.add_publication_from_pubmed}
\title{Higher-level fun to generate \code{add_publication_from_pubmed} util for one-off usage (default) or optimized for batch processing.}
\usage{
.add_publication_from_pubmed(batch = 0L, cache = batch)
}
\arguments{
\item{batch}{If a non-zero batch size, turns on batch mode; defaults to no-batch.}

\item{cache}{Whether to cache some results, which is default if \code{batch}.}
}
\description{
Higher-level fun to generate \code{add_publication_from_pubmed} util for one-off usage (default) or optimized for batch processing.
}
\keyword{internal}
