% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{.dict_to_list}
\alias{.dict_to_list}
\title{Convert a flat Python Dict to R list}
\usage{
.dict_to_list(dict)
}
\arguments{
\item{dict}{A flat Python Dict object.}
}
\description{
An internal function used to convert Annotations objects returned by \code{get_annotations}.
}
