% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nfcore_manifest.R
\name{.generate_rnaseq_samplesheet}
\alias{.generate_rnaseq_samplesheet}
\title{Generate rnaseq samplesheet format}
\usage{
.generate_rnaseq_samplesheet(metadata, synstage, strandedness, dataset_id)
}
\arguments{
\item{metadata}{Data frame with dataset metadata}

\item{synstage}{Logical for Synapse staging}

\item{strandedness}{Strandedness setting}

\item{dataset_id}{Synapse dataset ID for staging}
}
\value{
Data frame in rnaseq samplesheet format
}
\description{
Generate rnaseq samplesheet format
}
\keyword{internal}
