% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nfcore_manifest.R
\name{.generate_sarek_samplesheet}
\alias{.generate_sarek_samplesheet}
\title{Generate sarek samplesheet format}
\usage{
.generate_sarek_samplesheet(metadata, synstage, dataset_id)
}
\arguments{
\item{metadata}{Data frame with dataset metadata}

\item{synstage}{Logical for Synapse staging}

\item{dataset_id}{Synapse dataset ID for staging}
}
\value{
Data frame in sarek samplesheet format
}
\description{
Generate sarek samplesheet format
}
\keyword{internal}
