% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_stringlist.R
\name{.modify_annotation}
\alias{.modify_annotation}
\title{Modify a single annotation on a single file}
\usage{
.modify_annotation(synapse_id, key, value)
}
\arguments{
\item{synapse_id}{A synapse entity id.}

\item{key}{The key of the annotation to modify.}

\item{value}{The value to change the annotation to.}
}
\description{
Modifies a single annotation value on a single (existing) synapse file.
}
