% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pubmed_publication.R
\name{.store_publication}
\alias{.store_publication}
\title{Adds a row to a table.}
\usage{
.store_publication(schema, new_row)
}
\arguments{
\item{schema}{A synapse table Schema object.}

\item{new_row}{A data frame of one or more rows that match the provided schema.}
}
\description{
Adds a row to a table.
}
