% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_publication_from_unpaywall.R
\name{.store_rows}
\alias{.store_rows}
\title{Adds a row to a table.}
\usage{
.store_rows(schema, new_row)
}
\arguments{
\item{schema}{A synapse table Schema object.}

\item{new_row}{A data frame of one or more rows that match the provided schema.}
}
\description{
Adds a row to a table.
}
