% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_related_studies.R
\name{.update_table_data}
\alias{.update_table_data}
\title{Replace/update table contents = input data must have ROW_ID and ROW_VERSION columns to update, otherwise will append data.}
\usage{
.update_table_data(table_id, new_data, etag = NULL)
}
\arguments{
\item{table_id}{The synapse id of the table to update.}

\item{new_data}{The updated table.}

\item{etag}{An etag of the latest version of the table. If not provided, will query table_id to retrieve latest etag.}
}
\description{
Replace/update table contents = input data must have ROW_ID and ROW_VERSION columns to update, otherwise will append data.
}
