% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_study_annotations.R
\name{.update_view_data}
\alias{.update_view_data}
\title{Replace/update table contents = input data must have ROW_ID, ROW_VERSION, ETAG columns to update.}
\usage{
.update_view_data(table_id, new_data)
}
\arguments{
\item{table_id}{The synapse id of the table to update.}

\item{new_data}{The updated table.}
}
\description{
Replace/update table contents = input data must have ROW_ID, ROW_VERSION, ETAG columns to update.
}
