% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrams.R
\name{dsp_dataset_mapping}
\alias{dsp_dataset_mapping}
\title{Wrapper to create Data Sharing Plan to project dataset comparison chart}
\usage{
dsp_dataset_mapping(dsp_datasets, project_datasets)
}
\arguments{
\item{dsp_datasets}{Named vector of datasets in data sharing plan.}

\item{project_datasets}{Named vector of datasets in project.}
}
\description{
Wrapper to create Data Sharing Plan to project dataset comparison chart
}
