% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_utils.R
\name{dt_read}
\alias{dt_read}
\title{Download and read file to \code{data.table}}
\usage{
dt_read(file)
}
\description{
Convenience function for reading a delimited local file or one on Synapse.
}
\keyword{internal}
