% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_testing_utils.R
\name{extract_libtype}
\alias{extract_libtype}
\title{Extract the library type from multiqc_data.json in nf-core/rnaseq output}
\usage{
extract_libtype(multiqc_id)
}
\arguments{
\item{syn_id}{A synapse id for the multiqc_data.json file.}
}
\value{
a data frame of library types
}
\description{
Extract the library type from multiqc_data.json in nf-core/rnaseq output
}
