% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_annotation_utils.R
\name{extract_syn_id_from_ss}
\alias{extract_syn_id_from_ss}
\title{Extract Synapse id from URI}
\usage{
extract_syn_id_from_ss(x)
}
\description{
Given some vector \code{x} of URIs/file paths,
will try different methods to extract what's likely the needed Synapse id,
then sanity checks of results.
}
\seealso{
\code{\link[=bare_syn_id]{bare_syn_id()}}
}
\keyword{internal}
