% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_child}
\alias{find_child}
\title{Find id of a child entity in a container}
\usage{
find_child(child_name, parent)
}
\arguments{
\item{child_name}{Name of child entity.}

\item{parent}{Parent container (project or folder).}
}
\description{
Find id of a child entity in a container
}
