% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_child_type}
\alias{find_child_type}
\title{Find children of type}
\usage{
find_child_type(parent, child_type = list("file"))
}
\arguments{
\item{parent}{Parent container (project or folder).}

\item{child_type}{Type(s) as a list, even for only one type. Defaults to "file".}
}
\description{
Related to \code{find_child}, but instead of using a specific name, uses the type.
Returns a vector of ids, with entity names set as names.
}
