% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_data_root}
\alias{find_data_root}
\title{Find data folder}
\usage{
find_data_root(project_id)
}
\arguments{
\item{project_id}{Synapse project id.}
}
\description{
Convenience function to find data folder, which can have slight name variations, in a project.
}
