% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_in}
\alias{find_in}
\title{Find in path}
\usage{
find_in(scope, path)
}
\arguments{
\item{scope}{Id of the container (project or folder) to begin search.}

\item{path}{Path string in format "subdir1/subdir2/file.txt", where the last-level element will be the id returned.}
}
\description{
Get the Synapse id of an entity nested several folder layers deep without
having to click through the UI or create a fileview as long as the structure/path is known.
}
