% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_nf_asset}
\alias{find_nf_asset}
\title{Find a standard nextflow workflow output asset}
\usage{
find_nf_asset(
  syn_out,
  asset = c("software_versions", "multiqc_report", "samplesheet", "samtools_stats")
)
}
\arguments{
\item{syn_out}{Id of top-level folder that corresponds to \code{publishDir} in a nextflow workflow.}

\item{asset}{Name of asset to find.}
}
\value{
Id of samplesheet.
}
\description{
Note that samplesheets became part of the output only for newer versions of nf-core/rna-seq;
older runs may not find samplesheets.
Paths default to known working paths corresponding to the latest major workflow version,
but this may change and may need to be updated as part of util maintenance.
}
