% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_publication_from_pubmed.R
\name{from_pubmed}
\alias{from_pubmed}
\title{Get publication metadata from PubMed}
\usage{
from_pubmed(pmid)
}
\arguments{
\item{pmid}{PubMed id.}
}
\value{
If PMID found, return meta as table w/ \code{title} \code{journal} \code{author} \code{year} \code{pmid} \code{doi}.
}
\description{
Get publication metadata from PubMed
}
