% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remanifest.R
\name{gather_annotations}
\alias{gather_annotations}
\title{Internal helper for gathering annotations into a table using the REST API}
\usage{
gather_annotations(ids, list_sep = ", ")
}
\arguments{
\item{ids}{One or more ids.}

\item{list_sep}{List separator for list annotations.}
}
\description{
This is an internal implementation that works directly with the platform service JSON
to afford more low-level control and avoid Python-R object conversion differences between reticulate versions.
}
\keyword{internal}
