% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_utils.R
\name{get_by_prop_from_json_schema}
\alias{get_by_prop_from_json_schema}
\title{Look up connected nodes by specified property in JSON-LD schema}
\usage{
get_by_prop_from_json_schema(
  id,
  prop,
  schema =
    "https://raw.githubusercontent.com/nf-osi/nf-metadata-dictionary/main/NF.jsonld",
  return_labels = TRUE,
  recursive = FALSE,
  result = NULL,
  rest = NULL
)
}
\arguments{
\item{id}{Id (\verb{@id}) for which to get range values; include prefix if needed.}

\item{prop}{Property; include prefix if needed.}

\item{schema}{Path (URL or local) to file from which schema will be read, or schema as list object.}

\item{return_labels}{Return labels (default), otherwise ids of connected nodes.}

\item{recursive}{Recursive lookup?}

\item{result}{Vector of accumulated results; used for recursive lookup.}

\item{rest}{Vector of remaining ids; used for recursive lookup.}
}
\description{
Use with schematic-generated JSON-LD schema: given \verb{@id}, get connected nodes by specified prop (e.g. \code{sms:something}).
Intended to be a generic used to define more specific lookup utils.
Can do recursive lookup, though graph should be a tree/acyclic (!).
(Useful for props such as \code{dependsOn}, doesn't make sense for props such as \code{rdfs:label}.)
}
