% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncBP.R
\name{get_data_for_releasable}
\alias{get_data_for_releasable}
\title{Download data from a view for releasable samples in samplesheet}
\usage{
get_data_for_releasable(samplesheet, ref_view, verbose = TRUE)
}
\arguments{
\item{samplesheet}{Samplesheet \code{data.table}.}

\item{ref_view}{View to get data from.}

\item{verbose}{Output details.}
}
\description{
This tries to check that complete data could be retrieved from said view.
Note: Since the view is typically denormalized, not all data might be clinical.
A downstream step will do some of the additional processing/subsetting needed.
}
\keyword{internal}
