% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_utils.R
\name{get_dependency_from_json_schema}
\alias{get_dependency_from_json_schema}
\title{Get dependencies for node in JSON-LD schema}
\usage{
get_dependency_from_json_schema(
  id,
  prop = "sms:requiresDependency",
 
    schema = "https://raw.githubusercontent.com/nf-osi/nf-metadata-dictionary/main/NF.jsonld",
  return_labels = TRUE,
  recursive = TRUE,
  result = NULL,
  rest = NULL
)
}
\arguments{
\item{id}{Id (\verb{@id}) for which to get range values; include prefix if needed.}

\item{prop}{Property; include prefix if needed.}

\item{schema}{Path (URL or local) to file from which schema will be read, or schema as list object.}

\item{return_labels}{Return labels (default), otherwise ids of connected nodes.}

\item{recursive}{Recursive lookup?}

\item{result}{Vector of accumulated results; used for recursive lookup.}

\item{rest}{Vector of remaining ids; used for recursive lookup.}
}
\description{
Shorthand for getting props defined in annotation template using \code{get_by_prop_from_json_schema} under the hood.
}
