% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{get_doi_meta}
\alias{get_doi_meta}
\title{Get DOI metadata if it exists}
\usage{
get_doi_meta(id)
}
\arguments{
\item{id}{Dataset data.}
}
\description{
Returns list of metadata associated with DOI if exists, otherwise NULL.
Currently usable for certain entity types like files or datasets,
though this should be revised to make more useful with other objects.
Note: Internal/experimental use only, not for production use.
}
\keyword{internal}
