% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{get_gs_project_tracking}
\alias{get_gs_project_tracking}
\alias{get_project_tracking}
\title{Get and parse data from Google Sheets for initializing a new project}
\usage{
get_gs_project_tracking(
  sheet,
  creds = NULL,
  cols = c(name = "studyName", pi = "studyPI", lead = "studyLeads", admin_user = NULL,
    abstract = "abstract", institution = "institutions", funder = "fundingAgency",
    initiative = "initiative", disease = "diseaseFocus", manifestations =
    "diseaseManifestations", grant_doi = "grantDOI")
)

get_project_tracking(
  sheet,
  creds = NULL,
  cols = c(name = "studyName", pi = "studyPI", lead = "studyLeads", admin_user = NULL,
    abstract = "abstract", institution = "institutions", funder = "fundingAgency",
    initiative = "initiative", disease = "diseaseFocus", manifestations =
    "diseaseManifestations", grant_doi = "grantDOI")
)
}
\arguments{
\item{sheet}{Sheet URL or id. See \code{\link[googlesheets4]{read_sheet}}.}

\item{creds}{Path to JSON creds file (service account token).}

\item{cols}{List of columns that map to required parameters for \code{\link{new_project}}.
Defaults are provided.}
}
\description{
Notice: This is DEPRECATED and will be removed in the next version.
Currently, project tracking data is stored in a private GoogleSheet.
For \code{\link{new_project}}, this wraps \code{googlesheets4} to get the needed data.
}
\details{
If \code{creds} is not provided, i.e. there is no service account token,
then usage requires authenticating the Tidyverse API in the browser to get an API token on your behalf.
This doesn't actually give the googlesheets4 project access to any data
("The Tidyverse API Packages project never receives your data or the permission to access your data."
-- see https://www.tidyverse.org/google_privacy_policy/.)
}
\keyword{internal}
