% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{get_path}
\alias{get_path}
\title{Get path for a Synapse id}
\usage{
get_path(id, path = NULL)
}
\arguments{
\item{id}{Synapse id.}

\item{path}{Path builder value.}
}
\description{
Small helper fun to get path for given Synapse entity.
}
