% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_wiki_links.R
\name{get_project_wiki}
\alias{get_project_wiki}
\title{Get wiki content of synapse project(s)}
\usage{
get_project_wiki(project_id, markdown = TRUE)
}
\arguments{
\item{project_id}{Character vector of synapse project id(s) for which to get wiki.}

\item{markdown}{If \code{TRUE} (default) return only the markdown text, else return full wiki object.}
}
\value{
A list storing the wiki object or markdown-formatted text.
}
\description{
Get the wiki object or text content (main page only). This is primarily a helper function used in QC but may be useful for other wiki analysis.
}
\examples{
\dontrun{ 
txt <- get_project_wiki(c("syn11374354","syn2343195"))
}
}
