% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_study_data_types.R
\name{get_valid_values_from_json_schema}
\alias{get_valid_values_from_json_schema}
\title{Retrieve valid subclasses of a value in a JSON-LD schema}
\usage{
get_valid_values_from_json_schema(
 
    schema_url = "https://raw.githubusercontent.com/nf-osi/schematic/develop/data/schema_org_schemas/NF.jsonld",
  source_name_schema = "DataType"
)
}
\arguments{
\item{schema_url}{Default: the NF-OSI JSON-LD schema.}

\item{source_name_schema}{Default = DataType. The value for which you'd like to find the associated subclasses.}
}
\value{
A character vector of values.
}
\description{
Retrieve valid subclasses of a value in a JSON-LD schema. Notably, I know very little about JSON-LD, so this function is 1: very hacky, and 2: may not work on other JSON-LD files.
}
