% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_annotation.R
\name{get_workflow_ref}
\alias{get_workflow_ref}
\title{Translate output folders to workflow context specific metadata}
\usage{
get_workflow_ref(output_name, workflow_context)
}
\arguments{
\item{output_name}{Name of the output directory.}

\item{workflow_context}{Supported nextflow workflow context, "rnaseq" or "sarek".}
}
\description{
This is a manually maintained reference that covers nextflow outputs for
https://nf-co.re/sarek/{version}/docs/output/#qc-and-reporting & https://nf-co.re/rnaseq/{version}/docs/output/#quality-control,
where reference should be accurate for rnaseq version = 3.x.x and sarek version = 2.x.x - 3.x.x.
}
\keyword{internal}
