% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_testing_utils.R
\name{identify_read_pair}
\alias{identify_read_pair}
\title{Identify read pair from string}
\usage{
identify_read_pair(string)
}
\arguments{
\item{string}{A filename string.}
}
\value{
Returns a read pair: 1, 2, or NULL if none detected.
}
\description{
Identify read pair from string
}
