% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_qc.R
\name{infer_data_type}
\alias{infer_data_type}
\title{Infer data type of a dataset folder}
\usage{
infer_data_type(dataset_id)
}
\arguments{
\item{dataset_id}{Synapse ID of the dataset to query.}
}
\value{
\item{data_type}{Inferred data type of dataset_id. NA if unable to infer data 
type or conflicting data types observed.}
}
\description{
Infer the data type by checking the first few files.
}
