% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_utils.R
\name{inherit_input_annotations}
\alias{inherit_input_annotations}
\title{Inherit annotations}
\usage{
inherit_input_annotations(sample_io, select = NULL, update = FALSE)
}
\arguments{
\item{sample_io}{A \code{data.table} that minimally contains \code{input_id} and \code{output_id}.}

\item{select}{Vector of properties to selectively copy if present on the entity.
If not specified, will copy over everything, which may not be desirable.}

\item{update}{Whether to have annotations applied immediately or return manifest component only.}
}
\description{
Have output/derived entities inherit annotations from input entities.
If there are multiple inputs, inherit properties from the FIRST input.
Other options may be implemented in the future.
}
