% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{inherit_props}
\alias{inherit_props}
\title{Determine inherited properties}
\usage{
inherit_props(template, schema)
}
\arguments{
\item{template}{URI of data template in model, prefixed if needed.}

\item{schema}{Path (URL or local) to file from which schema will be read, or schema as list object.}
}
\description{
Internal function needed until there is better handling of property inheritance.
}
\keyword{internal}
