% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_utils.R
\name{is_valid_syn_id}
\alias{is_valid_syn_id}
\title{Validate a Synapse ID}
\usage{
is_valid_syn_id(id)
}
\arguments{
\item{id}{Id string.}
}
\description{
Returns the id if valid, throws error if not.
}
\keyword{internal}
