% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_qc.R
\name{list_project_datasets}
\alias{list_project_datasets}
\title{List datasets in project}
\usage{
list_project_datasets(project_id)
}
\arguments{
\item{project_id}{Synapse project id.}
}
\description{
Return a list of dataset folders if they are in expected location in project, otherwise NULL w/ explanatory message.
}
