% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_utils.R
\name{local_view}
\alias{local_view}
\title{Create local view of some scopes}
\usage{
local_view(scope, idcol = "parent", idmap = NULL)
}
\arguments{
\item{scope}{Vector of one or more syn container ids.}

\item{idcol}{Name of an idcol; useful to keep track of which rows came from which scope.
Can be NULL if not needed.}

\item{idmap}{Since id col will default to the syn id,
providing a map will allow tracking by container name instead.}
}
\description{
Creating a fileview or dataset can be useful for filtering/selection,
but sometimes one wants to avoid having to do that directly on synapse,
because queries and query result downloads can be slower and not as facile.
As well, it can be somewhat messy to have many helper fileviews in a project.
This creates a local view with default props that works well enough for certain jobs.
}
