% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{make_case_list_maf}
\alias{make_case_list_maf}
\title{Case lists for mutation samples}
\usage{
make_case_list_maf(cancer_study_identifier, verbose = TRUE)
}
\description{
https://docs.cbioportal.org/file-formats/#case-lists
}
\keyword{internal}
