% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{make_cbio_clinical_header}
\alias{make_cbio_clinical_header}
\title{Make header for cBioPortal clinical data file}
\usage{
make_cbio_clinical_header(df, mapping)
}
\arguments{
\item{df}{A \code{data.frame} representing clinical dataset to publicize.}

\item{mapping}{A reference table providing \code{label}, \code{description}, and \code{data_type} for each \code{source} attribute in \code{df}.}
}
\description{
This is called from the wrapper \code{write_cbio_clinical}.
Adapted from https://github.com/Sage-Bionetworks/genie-erbb2-cbio/blob/develop/create_clinical.R#L396.
Needs a data table of clinical data and a reference providing \code{label}, \code{description}, and \code{data_type}.
}
\keyword{internal}
