% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{make_meta_clinical_generic}
\alias{make_meta_clinical_generic}
\title{Generic template for clinical data file}
\usage{
make_meta_clinical_generic(
  cancer_study_identifier,
  genetic_alteration_type,
  datatype,
  data_filename
)
}
\arguments{
\item{cancer_study_identifier}{The study identifier.}

\item{genetic_alteration_type}{The cBioPortal generic alteration type.}

\item{datatype}{The cBioPortal data type of \code{data_filename}.}

\item{data_filename}{Name of the data file that this meta file describes.}
}
\description{
Make meta file to describe one of the clinical data files (e.g. SAMPLE, PATIENT).
Adapted from https://github.com/Sage-Bionetworks/genie-erbb2-cbio/blob/develop/make_meta.R#L65
}
\keyword{internal}
