% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{make_meta_cna}
\alias{make_meta_cna}
\title{Make meta file for cBioPortal copy number alteration data}
\usage{
make_meta_cna(
  cancer_study_identifier,
  data_filename = "data_cna.seg",
  reference_genome_id = "hg19",
  publish_dir = ".",
  write = TRUE,
  verbose = TRUE
)
}
\description{
Currently assumes seg data and should be extended later.
}
\details{
See https://docs.cbioportal.org/file-formats/#segmented-data
}
\keyword{internal}
