% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{make_meta_expression}
\alias{make_meta_expression}
\title{Make meta file for cBioPortal expression data}
\usage{
make_meta_expression(
  cancer_study_identifier,
  data_filename = "data_expression.txt",
  publish_dir = ".",
  write = TRUE,
  verbose = TRUE
)
}
\description{
https://docs.cbioportal.org/file-formats/#expression-data
}
\keyword{internal}
