% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{make_meta_genomic_generic}
\alias{make_meta_genomic_generic}
\title{Generic template for genomic-type data file}
\usage{
make_meta_genomic_generic(
  cancer_study_identifier,
  genetic_alteration_type,
  datatype,
  stable_id,
  profile_name,
  profile_description,
  data_filename
)
}
\arguments{
\item{cancer_study_identifier}{The study identifier.}

\item{genetic_alteration_type}{The cBioPortal generic alteration type.}

\item{datatype}{The cBioPortal data type of \code{data_filename}.}

\item{stable_id}{Stable id.}

\item{profile_name}{Name of the genomic profiling. This is set by the more specific \code{make_meta} utility.
For example, "Mutations" for \verb{make_*_maf} and "Copy-number alterations" for \verb{make_*_cna}.}

\item{profile_description}{Brief description for the genomic profiling.
This is set by the more specific \code{make_meta} utility.}

\item{data_filename}{Name of the data file that this meta file describes.}
}
\description{
Reused from https://github.com/Sage-Bionetworks/genie-erbb2-cbio/blob/develop/make_meta.R#L65
}
\keyword{internal}
