% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{make_meta_maf}
\alias{make_meta_maf}
\title{Make meta file for maf}
\usage{
make_meta_maf(
  cancer_study_identifier,
  data_filename = "data_mutations.txt",
  publish_dir = ".",
  write = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{cancer_study_identifier}{The study identifier.}

\item{data_filename}{Name of the data file. Defaults to "data_mutations.txt".}

\item{publish_dir}{Directory path to write to, defaults to current.}

\item{write}{Whether to write the meta file for the data file.}

\item{verbose}{Report where file has been written.}
}
\description{
Reused from https://github.com/Sage-Bionetworks/genie-erbb2-cbio/blob/develop/make_meta.R#L157
}
\keyword{internal}
