% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{make_meta_patient}
\alias{make_meta_patient}
\title{Make patient meta file}
\usage{
make_meta_patient(
  cancer_study_identifier,
  data_filename = "data_clinical_patient.txt",
  write = TRUE,
  publish_dir = ".",
  verbose = TRUE
)
}
\arguments{
\item{cancer_study_identifier}{The study identifier.}

\item{data_filename}{Name of the data file that this meta file describes.}

\item{write}{Whether to write the meta file for the clinical data file.}

\item{publish_dir}{Directory path to write to, defaults to current.}

\item{verbose}{Report where file has been written.}
}
\description{
Adapted from https://github.com/Sage-Bionetworks/genie-erbb2-cbio/blob/develop/create_meta.R#L101
}
\keyword{internal}
