% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{make_meta_study}
\alias{make_meta_study}
\title{Make meta study file}
\usage{
make_meta_study(
  cancer_study_identifier,
  type_of_cancer = "mixed",
  name,
  description,
  short_name,
  publish_dir = ".",
  write = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{cancer_study_identifier}{The study identifier.}

\item{type_of_cancer}{Type of cancer, defaults to "mixed". See also http://oncotree.mskcc.org/#/home.}

\item{name}{Name of the study.}

\item{description}{Description of the study.}

\item{short_name}{Short name for the study.}

\item{publish_dir}{Directory path to write to, defaults to current.}

\item{write}{Whether to write the meta file for the clinical data file.}

\item{verbose}{Report where file has been written.}
}
\description{
Reused from https://github.com/Sage-Bionetworks/genie-erbb2-cbio/blob/develop/create_meta.R#L179
}
