% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{make_meta_study_generic}
\alias{make_meta_study_generic}
\title{Template for meta study file}
\usage{
make_meta_study_generic(
  type_of_cancer,
  cancer_study_identifier,
  name,
  description,
  groups,
  short_name
)
}
\description{
Template for meta study file
}
\keyword{internal}
