% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_utils.R
\name{make_public}
\alias{make_public}
\title{Make public}
\usage{
make_public(id)
}
\arguments{
\item{id}{Synapse entity id.}
}
\description{
Sets READ/DOWNLOAD permissions for web and registered users equivalently to the "Make Public" button in Synapse UI.
TODO: For regular users this can be a one-and-done action, but for the DCC admin this likely entails some other actions,
such as updating a project tracking table, so a wrapper or "callback" functionality might be needed.
}
