% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_utils.R
\name{make_public_viewable}
\alias{make_public_viewable}
\title{Set public access to VIEW (READ) only for an entity}
\usage{
make_public_viewable(id)
}
\arguments{
\item{id}{Synapse entity id.}
}
\description{
Set both registered users and non-registered users to have VIEW-only permissions.
See code{link{make_public}} for more permissive permissions to download (for registered users), which is usually set later at data release time.
}
