% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_qc.R
\name{manifest_generate}
\alias{manifest_generate}
\title{Generate manifest via schematic service}
\usage{
manifest_generate(
  data_type,
  dataset_id = NULL,
  title = data_type,
  schema_url =
    "https://raw.githubusercontent.com/nf-osi/nf-metadata-dictionary/main/NF.jsonld",
  asset_view = "syn16858331",
  output_format = "excel",
  use_annotations = TRUE,
  service = "https://schematic.api.sagebionetworks.org/v1/manifest/generate"
)
}
\arguments{
\item{data_type}{Data type of the manifest to generate (aka Component).}

\item{dataset_id}{Optional, if given this fills out manifest for existing dataset instead of generating a blank manifest.}

\item{title}{Optional, custom title.}

\item{schema_url}{Optional, defaults to main NF 'latest' data model.}

\item{asset_view}{Optional, defaults to main NF portal fileview.}

\item{output_format}{Format of 'excel', 'google_sheet', or 'dataframe'. Defaults to 'excel'.}

\item{use_annotations}{Use annotations if filling out manifest for existing dataset. Defaults to TRUE for NF.}

\item{service}{Service endpoint to use. Defaults to the schematic production endpoint.}
}
\value{
For excel, path to local file; for google_sheet, URL to sheet; for dataframe, JSON string of data.
}
\description{
See \href{https://schematic.api.sagebionetworks.org/v1/ui/#/Manifest\%20Operations/schematic_api.api.routes.get_manifest_route}{schematic manifest generation}.
Note that this uses the access token of user that should already be logged in with \code{syn_login}.
}
