% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_qc.R
\name{manifest_passed}
\alias{manifest_passed}
\title{Provide a pass/fail summary result}
\usage{
manifest_passed(result)
}
\arguments{
\item{result}{Result list data from schematic service.}
}
\value{
Boolean for whether passed.

List of structure \code{list(result = result, notes = notes)}, where \code{result} indicates whether the dataset passed.
}
\description{
Provide a pass/fail summary result
}
\keyword{internal}
