% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_qc.R
\name{manifest_validate}
\alias{manifest_validate}
\title{Validate manifest via schematic service}
\usage{
manifest_validate(
  data_type,
  json_str = NULL,
  file_name = NULL,
  restrict_rules = FALSE,
  schema_url =
    "https://raw.githubusercontent.com/nf-osi/nf-metadata-dictionary/main/NF.jsonld",
  service = "https://schematic.api.sagebionetworks.org/v1/model/validate"
)
}
\arguments{
\item{data_type}{Data type of the manifest to generate (aka Component).}

\item{json_str}{JSON string representing metadata.}

\item{file_name}{Path to file, has to be \code{.csv}. Ignored if \code{json_str} is given.}

\item{restrict_rules}{Use only basic schematic validation instead of extended validation with Great Expectations, default \code{FALSE}.}

\item{schema_url}{Optional, defaults to main NF 'latest' data model.}

\item{service}{Service endpoint to use. Defaults to the schematic production endpoint.}
}
\description{
See \href{https://schematic.api.sagebionetworks.org/v1/ui/#/Model\%20Operations/schematic_api.api.routes.validate_manifest_route}{schematic validation}.
Get validation results from schematic service. Downstream utils can consume these results for custom display/report.
}
