% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_qc.R
\name{manifest_validate_wrapper}
\alias{manifest_validate_wrapper}
\title{Validate with stated data_type in manifest}
\usage{
manifest_validate_wrapper(
  csv_file,
  data_type = NULL,
  dataset_id = NULL,
  dataset_name = NULL
)
}
\arguments{
\item{csv_file}{Path to the manifest csv file.}

\item{data_type}{Optional if present in manifest.}

\item{dataset_id}{Optional dataset id.}

\item{dataset_name}{Optional dataset name.}
}
\description{
Validate with stated data_type in manifest
}
