% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_utils.R
\name{map_sample_input}
\alias{map_sample_input}
\title{Parse a nextflow samplesheet to construct input-output map}
\usage{
map_sample_input(samplesheet, parse_fun = function(x) gsub("_T[0-9]$", "", x))
}
\arguments{
\item{samplesheet}{A local file or syn id of samplesheet.}

\item{parse_fun}{Function implementing how to parse samples in samplesheet.}
}
\description{
Samplesheets are used in rnaseq pipelines,
and are defined here: https://nf-co.re/rnaseq/usage#full-samplesheet.
After the pipeline is run, it should be found in an output folder called \code{pipeline_info}.
}
\details{
Currently this provides refs for provenance annotation by returning
a mapping of sample names to input files (usually one-to-many but can be one-to-one),
which can be passed to downstream helpers either as a table
(which does not differ much from the samplesheet) or as a list.
}
