% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_annotation_utils.R
\name{map_sample_io}
\alias{map_sample_io}
\title{Map sample input-output}
\usage{
map_sample_io(input, output)
}
\arguments{
\item{input}{Data from \code{map_sample_input_ss}.}

\item{output}{Data from \verb{map_sample_output_*}.}
}
\value{
A table that includes \code{sample} \code{level} \code{output_id} \code{output_name} \code{input_id}.
}
\description{
Wrapper to map sample inputs and outputs depending on workflow type.
}
