% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_utils.R
\name{map_sample_output_rnaseq}
\alias{map_sample_output_rnaseq}
\title{Map sample to output from nf-rnaseq}
\usage{
map_sample_output_rnaseq(syn_out)
}
\arguments{
\item{syn_out}{Syn id of syn output destination (folder) with files of interest.}
}
\description{
See https://nf-co.re/rnaseq.
In specified location where workflow has deposited the outputs (i.e. "result" directory),
map out the relevant processed files based on file extensions and tie these files to source samples.
This relies on expected file locations and file naming convention of the specific workflow.
WARNING: Be aware that this may not work using a different version of workflow
or if files are manually re-organized post-write to Synapse.
With the known Nextflow workflow, outputs are organized by sample-id folders,
e.g. \verb{results/salmon/<sample>/<file>}, so this looks for processed files within one level of nesting.
}
\details{
See the related \code{\link{map_sample_input_ss}} for mapping sample to inputs instead of outputs.
}
