% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_fix.R
\name{match_col}
\alias{match_col}
\title{Find matching col in schema based on name}
\usage{
match_col(schema, col_name)
}
\description{
Synapse doesn't allow schemas to have columns of same name,
so this should never return more than one.
}
\keyword{internal}
