% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{match_exactly_one}
\alias{match_exactly_one}
\title{Helper to check for and get exactly one matching target element with given data}
\usage{
match_exactly_one(target, df, m)
}
\description{
Helper to check for and get exactly one matching target element with given data
}
\keyword{internal}
