% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncBP.R
\name{match_maf_sample_id}
\alias{match_maf_sample_id}
\title{Match clinical data with maf sample ids}
\usage{
match_maf_sample_id(clinical_data, merged_maf = NULL)
}
\arguments{
\item{clinical_data}{Clinical data as a \code{data.table}.}

\item{merged_maf}{Maf data as a \code{data.table}.}
}
\description{
PATIENT_ID and SAMPLE_ID can only contain letters, numbers, points, underscores and/or hyphens.
In the nf processing, sample id spaces are replaced with underscores in the \code{maf},
so this is applied to clinical data to match.
}
\keyword{internal}
