% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_qc.R
\name{meta_qc_dataset}
\alias{meta_qc_dataset}
\title{QC dataset metadata with pass/fail result}
\usage{
meta_qc_dataset(
  dataset_id,
  data_type = NULL,
  asset_view = "syn16858331",
  schema_url =
    "https://raw.githubusercontent.com/nf-osi/nf-metadata-dictionary/main/NF.jsonld",
  cleanup = TRUE,
  depth = 1L
)
}
\arguments{
\item{dataset_id}{Id of folder that represents a dataset, not actual Synapse dataset entity -- see details.}

\item{data_type}{A specific data type to validate against, otherwise tries to infer based on annotations. See details.}

\item{asset_view}{A reference view, defaults to the main NF portal fileview.}

\item{schema_url}{Schema URL, points by default to 'latest' main NF schema, can change to use a specific released version.}

\item{cleanup}{Whether to automatically remove reconstituted manifests once done. Default \code{TRUE}.}

\item{depth}{How much deeper to go when there appears to be no files in the immediate scope. Defaults to 1L.}
}
\value{
List of structure \code{list(result = result, notes = notes)},
where \code{result} indicates passing or \code{NA} if no data or if couldn't be validated for other reasons.
}
\description{
R wrapper for validation workflow with schematic. Because there is no validation-only service endpoint,
we move metadata around twice (generating manifest from server and submitting back to server),
so once schematic has a validation-only service endpoint that would be much more efficient.
A dataset in this context is a folder, usually tagged with \code{contentType} = "dataset".
}
\details{
Note that we prefer to wrap the schematic web API over a local installation because:
\itemize{
\item Will not require user to go through local schematic setup for this to be functional
\item API more likely reflects an up-to-date version of schematic and consistent with current DCA deployment
}

When \code{data_type} can't be inferred based on annotations, this is treated as a fail.

Status: alpha and likely to change based on changes in \code{schematic}.
}
