% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_qc.R
\name{meta_qc_project}
\alias{meta_qc_project}
\title{QC metadata at the project level with pass/fail result}
\usage{
meta_qc_project(project_id, result_file = NULL, ...)
}
\arguments{
\item{project_id}{Synapse project id.}

\item{result_file}{If not NULL, \emph{also} write to output to \code{.csv} file.}

\item{...}{Params passed to \code{meta_qc_dataset}.}
}
\value{
A table of with rows for the datasets QC'd, with dataset id, name, TRUE/FALSE pass result, and summary;
otherwise \code{NA}.
}
\description{
Wrapper to go through project datasets and do revalidation in one-stop-shop manner.
This works best when datasets are directly under "Raw Data" (the standard and most-preferred organization)
\emph{or} at most one additional level of nesting. See https://help.nf.synapse.org/NFdocs/how-to-organize-data.
For selective validation or more complicated structures,
look at \code{meta_qc_dataset} to do manual or interactive dataset-by-dataset validation.
}
