% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_annotations_email.R
\name{missing_annotation_email}
\alias{missing_annotation_email}
\title{Convert a delimited string to a stringlist annotation}
\usage{
missing_annotation_email(
  fileview_id,
  annotation_key,
  created_date,
  dry_run = TRUE
)
}
\arguments{
\item{fileview_id}{The synapse id of a fileview. Must have the desired annotations in the schema, and must have the files to annotate included in the scope. Must have write access to the files you want to re-annotate.}

\item{annotation_key}{A character string of the annotation you'd like to use to detect unannotated files.}

\item{created_date}{The date ('DD/MM/YYYY') to cut off}

\item{dry_run}{Default = TRUE. Skips emailing and instead prints summary tibble.}
}
\value{
If dry_run == T, returns study tibble and skips upload.
}
\description{
Converts a delimited string to a stringlist annotation and adjust the associated schema in the portal fileview.
}
