% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial.R
\name{mock_files}
\alias{mock_files}
\title{Create n temp files}
\usage{
mock_files(n)
}
\arguments{
\item{n}{Integer number of files.}
}
\value{
Paths to files in the temp directory.
}
\description{
Create some text files for upload
}
