% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{new_dataset}
\alias{new_dataset}
\title{Create new dataset with given items}
\usage{
new_dataset(
  name,
  parent,
  items = NULL,
  item_version = NULL,
  folders = NULL,
  add_annotation_columns = TRUE,
  dry_run = TRUE
)
}
\arguments{
\item{name}{Name of the dataset. It should be unique within the \code{parent} project.}

\item{parent}{Synapse id of parent project where the dataset will live.}

\item{items}{A list of id(s) of items to include for adding items \emph{a la carte} to dataset.
Usually the same parent project storing the files, but in some cases it may be a different project.}

\item{item_version}{Integer for version that will be used for all items, e.g. 1.
Otherwise, "latest" or "stable_latest". See details.}

\item{folders}{A list of folder ids for adding all items in some folder(s). Can be used with \code{items}.}

\item{add_annotation_columns}{Whether to add annotation columns. Defaults to \code{TRUE}, same as the Python API.}

\item{dry_run}{If TRUE, don't actually store dataset, just return the data object for inspection or further modification.}
}
\description{
Create new dataset with given items
}
