% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{new_project}
\alias{new_project}
\title{Create a new project}
\usage{
new_project(
  name,
  pi,
  lead,
  admin_user = NULL,
  abstract,
  institution,
  funder = c("CTF", "GFF", "NTAP"),
  initiative,
  webview = FALSE
)
}
\arguments{
\item{name}{Name of the project/study.}

\item{pi}{Name of the principal investigator.}

\item{lead}{Name(s) of the project lead/data coordinator, comma-sep if multiple, e.g. "Jane Doe, John Doe".}

\item{admin_user}{(Optional) Synapse username of specified user to be made admin.
Currently, only takes one admin user, and the rest can be added via the UI.}

\item{abstract}{Project abstract/description.}

\item{institution}{Affiliated institution(s), \strong{semicolon-sep if multiple}, e.g. "Stanford University; University of California, San Francisco".}

\item{funder}{The funder org, currently one of c("CTF", "GFF", "NTAP"). The relevant funder team will be made admin.}

\item{initiative}{Title of funding initiative, e.g. "Young Investigator Award".}

\item{webview}{Whether to open web browser to view newly created project. Defaults to FALSE.}
}
\value{
The project object.
}
\description{
Set up a new NF project with wiki, folders, fileview, and permissions.
Most parameters come from a project intake & data sharing plan (DSP) form.
Aside from default folders, folders are tailored for data mentioned in DSP.
The NF-OSI team is hard-coded to be admin in addition to the funder team indicated by \code{funder}.
Since this is intended for actual new projects, it will fail if a same existing project is detected.
}
\details{
After project is created, NF Portal representation requires registration in backend:
\itemize{
\item New study row added to the Portal - Studies table.
\item Project added to Portal - Files scope.
}
}
