% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{new_project_strict}
\alias{new_project_strict}
\title{Create a strictly new project}
\usage{
new_project_strict(project_name)
}
\arguments{
\item{project_name}{Name of project to be created.}
}
\description{
Internal handler for creating a project that
first checks whether project already exists and disallows overwriting.
For a less strict version that allows overwriting with a warning,
e.g. named \code{update_project}, implement with
\code{createOrUpdate = TRUE} and then compare createdOn and modifiedOn to issue a warning
(which would be more informative than current Python client).
}
\keyword{internal}
