% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_nf.R
\name{nf_cnv_dataset}
\alias{nf_cnv_dataset}
\title{Create dataset for CNVKit results}
\usage{
nf_cnv_dataset(syn_out, parent, dry_run = TRUE)
}
\arguments{
\item{syn_out}{Output folder called 'cnvkit'}

\item{parent}{Synapse id of parent project where the dataset will live.}

\item{dry_run}{If TRUE, don't actually store dataset, just return the data object for inspection or further modification.}
}
\description{
Create dataset from all files in CNVKit output
}
