% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{nf_star_salmon_datasets}
\alias{nf_star_salmon_datasets}
\title{Create dataset for STAR-Salmon expression quantification results}
\usage{
nf_star_salmon_datasets(manifest, parent, dry_run = TRUE)
}
\arguments{
\item{manifest}{A table of annotated data manifest from \code{annotate_expression}.}

\item{parent}{Synapse id of parent project where the dataset will live.}

\item{dry_run}{If TRUE, don't actually store dataset, just return the data object for inspection or further modification.}
}
\description{
With a level-3 manifest that is created from \code{annotate_expression},
calls \code{new_dataset} to make quantification files (.sf) into dataset.
Uses latest version of the files and creates a "Draft" dataset.
See \code{nf_sarek_datasets}.
}
