% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{nf_workflow_version}
\alias{nf_workflow_version}
\title{Return workflow version according to workflow meta}
\usage{
nf_workflow_version(syn_out)
}
\arguments{
\item{syn_out}{Id of top-level folder that corresponds to \code{publishDir} in a nextflow workflow.}
}
\value{
Version string.
}
\description{
Return workflow version according to workflow meta
}
