% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrams.R
\name{ppp_mmd_template}
\alias{ppp_mmd_template}
\title{Pretty processing provenance mermaid template for the portal}
\usage{
ppp_mmd_template(
  project_nodes,
  input_dataset_nodes,
  folder_nodes,
  output_dataset_nodes,
  workflow_links,
  dataset_links,
  clicks
)
}
\description{
Mermaid flowchart helps navigate processed data with interactive links by default.
See also https://github.com/mermaid-js/mermaid.
}
\keyword{internal}
