% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_qc.R
\name{precheck_manifest}
\alias{precheck_manifest}
\title{Precheck a manifest}
\usage{
precheck_manifest(
  manifest_csv,
  official_props =
    "https://raw.githubusercontent.com/nf-osi/nf-metadata-dictionary/main/modules/props.yaml"
)
}
\arguments{
\item{manifest_csv}{Path to manifest_csv.}

\item{official_props}{(Optional) Doc listing official model attributes. Currently this requires the LinkML format.}
}
\description{
Precheck before sending manifest off to schematic validation service.
Provides additional context and helpful recommendations.
}
