% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal_charts.R
\name{processing_flowchart}
\alias{processing_flowchart}
\title{Wrapper to create data-driven flowchart with pretty processing provenance mermaid template}
\usage{
processing_flowchart(year = 1)
}
\arguments{
\item{year}{Year to determine subset of data for which to generate flowchart.}
}
\description{
Wrapper to create data-driven flowchart with pretty processing provenance mermaid template
}
\examples{
\dontrun{
flowchart <- processing_flowchart(year = 1)
cat(flowchart, file = "flowchart.mmd")
}
}
