% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_study.R
\name{register_study_files}
\alias{register_study_files}
\title{Add new project scope to "Portal - Files" fileview}
\usage{
register_study_files(project_id, portal_fileview = "syn16858331")
}
\arguments{
\item{project_id}{The project id, i.e. container, that will be added to the scope of the view.}

\item{portal_fileview}{Synapse id of "Portal - Files" entity view.}
}
\description{
Add a new project to the scope of the "Portal - Files" fileview so that
files for that project are "registered" and surfaced in portal.
}
