% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remanifest.R
\name{remanifest}
\alias{remanifest}
\title{Reconstitute a manifest}
\usage{
remanifest(scope, file = "manifest.csv", recurse = TRUE, mf = data.table())
}
\arguments{
\item{scope}{Synapse id of something representing one or more datasets. Can be a folder \emph{or} Synapse dataset entity.}

\item{file}{Name of file to be written. Use \code{NULL} to return the \code{data.table} instead of writing to file.}

\item{recurse}{Include everything nested within folder scope, default \code{TRUE}, or only the immediate files.
Not used if dataset is Synapse dataset entity.}

\item{mf}{Intermediate manifest to build upon; used when recursive.}
}
\description{
Quickly "reconstitute" a manifest for files with existing annotations.
This is a \emph{power-user} alternative to schematic's \code{manifest_generate} with option \code{useAnnotations = TRUE} for certain scenarios:
}
\details{
\itemize{
\item Meant to be faster and more convenient for workflows focused on scalable/programmatic annotation/revalidation.
Convenience and speed mainly comes from bypassing the schematic API layer and using the Synapse API directly,
so manifest is reconstituted directly as R \code{data.table} \strong{instead of having to download a Googlesheets or Excel intermediate}.
\strong{The result \code{.csv} is still intended to be schematic-compatible for validation via \code{manifest_validate}}.
\item Works with both a Synapse dataset entity as well as a folder;
schematic currently only handles the latter (the folder concept predates Synapse datasets).
\item If folder-scoped, can choose between generating manifest with just immediate files \emph{or}, like schematic, nested files.
This is useful, for example, if the main scope folder is "RNA-seq" and the main files are fastqs, but there is a child folder called "md5" with md5 files,
and the manifest is really meant for the fastqs.
\item Does not try to download \code{synapse_manifest.csv}, which can be useful to regenerate a manifest when ARs have been set.
\item Does not require an asset view.
}

Other developer notes:
See also \code{gather_annotations} for the under-the-hood workhorse.
Intended for further development to be used as part of advanced annotation utils.
}
