% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_mod.R
\name{remove_button}
\alias{remove_button}
\title{Remove button from a project wiki}
\usage{
remove_button(wiki, label, dry_run = TRUE)
}
\arguments{
\item{wiki}{The wiki object to operate on.}

\item{label}{Button label text.}

\item{dry_run}{Whether to return a wiki object only without actually performing update.}
}
\description{
This provides a way to remove buttons that should no longer be present, possibly as decided by newer and wiser design decisions.
See also \code{\link{button_widget}}.
The target button is selected based on the specified text label.
If for some reason there are multiple buttons with the same label, all of them will be removed.
}
