% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_mod.R
\name{remove_wiki_subpage}
\alias{remove_wiki_subpage}
\title{Remove a subpage from a project wiki}
\usage{
remove_wiki_subpage(project_id, subpage)
}
\arguments{
\item{project_id}{ID of the owner Synapse project.}

\item{subpage}{Name of the subpage}
}
\description{
Removes a wiki subpage by name (header).
Currently, this will decline to make any mods if there is not exactly one match for the subpage.
If there are multiple subpages of same name, it's not clear which is the right one to remove.
}
