% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{schema_max_str_len}
\alias{schema_max_str_len}
\title{Consult schema about max string length}
\usage{
schema_max_str_len(
  key,
 
    schema_url = "https://raw.githubusercontent.com/nf-osi/nf-metadata-dictionary/main/NF.jsonld",
  parent_context = "bts",
  default = 100
)
}
\description{
Utility to query the schema regarding the max string length for a key
based on current valid values. If the key values are not constrained (free-text),
a default string length of 100 is returned. If not in schema, returns NA.
TO DO: related fun to consult schema about type (integer, string, stringlist, etc.)
}
