% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{set_annotations}
\alias{set_annotations}
\title{Wrapper around the Python \code{set_annotations} that pulls current annotations
and adds new annotations with given annotations data or replaces
data for annotations with the same keys existing on the entity.}
\usage{
set_annotations(id, annotations)
}
\arguments{
\item{id}{Synapse entity id.}

\item{annotations}{A flat list representing annotation key-value pairs,
e.g. \code{list(foo = "bar", rank = 1, authors = c("jack", "jane"))}}
}
\description{
Wrapper around the Python \code{set_annotations} that pulls current annotations
and adds new annotations with given annotations data or replaces
data for annotations with the same keys existing on the entity.
}
