% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_study_data_types.R
\name{summarize_attribute}
\alias{summarize_attribute}
\title{Helper summarization util}
\usage{
summarize_attribute(
  summary_query,
  attribute,
  entity_id = NULL,
  dry_run = TRUE,
  check_fun = NULL
)
}
\arguments{
\item{summary_query}{Query (usually of a fileview) that returns appropriate aggregation per row. You may need to add \code{group_concat}, \code{distinct}, and or \code{unnest} to the query to get the correct list of distinct values depending on your data (e.g.\verb{select group_concat(distinct unnest(tumorType)) as tumorType from ...}).}

\item{attribute}{Name of attribute to update as annotation.}

\item{entity_id}{Either a single valid Synapse id of the entity for which to update the attribute \emph{or} a column present in \code{summary_query} that stores ids.}

\item{dry_run}{Default = \code{TRUE}. Whether to update as well or just return list of annotation objects.}

\item{check_fun}{An optional custom check function to apply to the values being updated in order for update to go through. Should return a boolean. Used only if dry_run = \code{FALSE}.
It can be tailored towards the attribute/entity being updated (i.e. taking into account the schema and valid values).}
}
\description{
Given some table X that provides values, summarize the values for an attribute and add summary as annotations on some entity.
The entity could be X itself or another entity Y, e.g. a parent container entity.
Example 1: With datasets, summarize \code{species} for all the files.
Example 2: With projects, summarize \code{dataType} for all the files (in fact, see \code{assign_study_data_types}).
}
