% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_utils.R
\name{summarize_file_access}
\alias{summarize_file_access}
\title{Summarize file access for files within some view}
\usage{
summarize_file_access(principal_id, access_type, fileview_id)
}
\arguments{
\item{principal_id}{Group(s) for which to check the access type.}

\item{access_type}{Which access type(s) to check for; result summarizes whether there are permissions for \emph{all} types specified.}

\item{fileview_id}{Syn id of the view. View must include \code{benefactorId} and \code{type}.}
}
\description{
Some portion adapted from \verb{@allaway}'s original script.
Common usages are:
\itemize{
\item Check numbers of files are viewable and downloadable for registered Synapse users:
\code{summarize_file_access(principal_id = 273948, access_type = "DOWNLOAD", "syn16858331")}
\item Check that files within the portal purview are actually editable for the NF-OSI Sage Team,
because we need at least edit permissions for updating annotations (historically, some issues with hackathon projects, etc.):
\code{summarize_file_access(principal_id = 3378999, access_type = "UPDATE", "syn16858331")}
}
}
\details{
More complex usage:
\itemize{
\item Check access for \emph{multiple} teams.
}

This summarizes file access by using the smaller set of benefactors,
and returns a \code{data.table} with columns \code{benefactorId}, \code{principalId}, \code{access}, and \code{N}, which describes whether
\code{principalId} has \emph{all} the specified \code{access} to \code{benefactorId} and its \code{N} files.
This data can be used for further aggregation as needed.
}
