% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_fix.R
\name{swap_col}
\alias{swap_col}
\title{Swap out old column for a new column in a schema}
\usage{
swap_col(schema, old, new)
}
\arguments{
\item{schema}{A table schema.}

\item{old}{Id of old col.}

\item{new}{Id of new col}
}
\description{
Swap out old column for a new column in a schema
}
\keyword{internal}
