% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{table_query}
\alias{table_query}
\title{Generic table query}
\usage{
table_query(table_id, columns = "*", includeRowIdAndRowVersion = F)
}
\arguments{
\item{table_id}{Synapse table id.}

\item{columns}{A character vector of selected columns (which often correspond to annotations, but not always).
If not given, will select all columns.}

\item{includeRowIdAndRowVersion}{Whether to include the row id and etag, defaults to \code{FALSE}. If the use case is to update rows in the table (rather than just to retrieve information for viewing, use \code{TRUE}).}
}
\description{
Retrieve selected data from a Synapse table.
}
