% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_qc.R
\name{tersely}
\alias{tersely}
\title{Terse error messages please}
\usage{
tersely(error)
}
\arguments{
\item{error}{An error object from schematic.}
}
\description{
Terse error messages please
}
\keyword{internal}
