% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_testing_utils.R
\name{test_failed}
\alias{test_failed}
\title{Check output strandedness matches samplesheet strandedness
TODO
Format a test fail message.}
\usage{
test_failed(display_string)
}
\arguments{
\item{display_string}{A character string to format as a test failed message.}
}
\value{
A message to the console
}
\description{
Check output strandedness matches samplesheet strandedness
TODO
Format a test fail message.
}
