% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_testing_utils.R
\name{test_passed}
\alias{test_passed}
\title{Format a test passed message.}
\usage{
test_passed(display_string)
}
\arguments{
\item{display_string}{A character string to format as a test passed message.}
}
\value{
A message to the console
}
\description{
Format a test passed message.
}
