% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{update_items}
\alias{update_items}
\title{INTERNAL - apply updates to a collection of items}
\usage{
update_items(current, update)
}
\arguments{
\item{current}{List of lists representing a collection of items.}

\item{update}{Collection of items to apply as updates to \code{current_items}.}
}
\description{
An internal transaction helper for trying to apply a changeset to a collection,
used in several higher-level collection utils.
Given the changeset that can represent updates of both types "replace" or "add",
this applies an update join keyed on \code{entityId} for the replace and
appends the new items to get the updated collection.
}
\keyword{internal}
