% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_study_annotations.R
\name{update_study_annotations}
\alias{update_study_annotations}
\title{Updates a set of files with project-level annotations.}
\usage{
update_study_annotations(study_table_id, fileview_id, annotations, dry_run = T)
}
\arguments{
\item{study_table_id}{The synapse id of the portal study table.}

\item{fileview_id}{The synapse id of a fileview. Must have the desired annotations in the schema, and must have the files to annotate included in the scope. Must have write access to the files you want to re-annotate.}

\item{annotations}{A vector of annotations to gather from the study table, and assign to the files.}

\item{dry_run}{Default = TRUE Skips upload of annotations unless set to FALSE.}
}
\value{
If dry_run == T, returns updated annotations as a tibble.
}
\description{
Adds studyId, studyName, initiative, funder annotations to a set of files from a fileview query. Assumes that projectId==studyId. You must have the Synapse default view column "type" present in your view. After running this function, changes may take a few minutes to propagate through Synapse.
}
\examples{
\dontrun{
update_study_annotations(study_table_id = "syn16787123",
                         fileview_id = "syn16858331",
                         annotations = c("studyId","studyName","initiative","fundingAgency"),
                         dry_run = T)
}
}
