% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{use_ref_map}
\alias{use_ref_map}
\title{Read and use a mapping file}
\usage{
use_ref_map(ref_map, as_dt = TRUE)
}
\arguments{
\item{ref_map}{YAML or JSON mapping. See details.}

\item{as_dt}{Return as \code{data.table}, the default,
otherwise do checking but just return the list representation, which retains some metadata.}
}
\value{
Either a list of lists storing \code{source}, \code{label}, \code{description}, \code{data_type}, \code{attribute_type}
or a \code{data.table} representation.
}
\description{
The mapping file should be a YAML or JSON format file that minimally has a mapping key
storing the translations of source data model element to the target data model.
See \href{https://github.com/nf-osi/nf-metadata-dictionary/tree/main/mappings}{example file for NF}.
This util reads the mapping file, does some light checking, and creates a list object that downstream functions can use.
}
\details{
If the specification for the mapping file is later formalized with a different structure
this should be handled accordingly (that is to say, this is one of the under-the-hood things that can change).
}
\keyword{internal}
