% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_bound_entity}
\alias{validate_bound_entity}
\title{Validate a \strong{schema-bound} entity}
\usage{
validate_bound_entity(id)
}
\arguments{
\item{id}{Entity id.}
}
\description{
This will error for unbound entities.
}
