% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_utils.R
\name{walk}
\alias{walk}
\title{Walk through a directory}
\usage{
walk(syn_id, as_list = TRUE)
}
\arguments{
\item{syn_id}{Synapse id of directory root to traverse.}
}
\value{
An R list or Py generator object.
}
\description{
For now, an internal util imported from \code{synapseutils}.
}
\keyword{internal}
