% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_mod.R
\name{wiki_mod}
\alias{wiki_mod}
\title{Add markup to a project wiki}
\usage{
wiki_mod(
  content,
  project_id,
  subpage = NULL,
  where = c("top", "bottom"),
  dry_run = TRUE
)
}
\arguments{
\item{content}{Markdown or other markup compatible with Synapse wikis.}

\item{project_id}{ID of the owner Synapse project.}

\item{subpage}{If given a character name, will add to a new subpage of that name.
If \code{NULL}, contents will be added to the main page.}

\item{where}{Where to add markup on page, "top" or "bottom" (defaults to "top").
Only used if adding to main page, which may already have content.}

\item{dry_run}{Whether to return a wiki object only without actually performing update.}
}
\description{
Add markup to an \strong{existing} project wiki, e.g. regular markdown, a widget, or other Synapse wiki compatible content.
Errors will be encountered if one tries to modify a project wiki that does not exist.
}
