% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{with_selected_elements}
\alias{with_selected_elements}
\title{Subset with selected data elements}
\usage{
with_selected_elements(df, selectable_elements)
}
\arguments{
\item{df}{Data.}

\item{selectable_elements}{Selectable elements.}
}
\description{
Warns if some elements present will be excluded.
}
\keyword{internal}
