% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{write_meta}
\alias{write_meta}
\title{Write meta file}
\usage{
write_meta(data, filename, publish_dir = ".", verbose = TRUE)
}
\arguments{
\item{data}{The data (lines) to write.}

\item{filename}{Name of file.}

\item{publish_dir}{Directory path to write to, defaults to current.}

\item{verbose}{Report where file has been written.}
}
\description{
Slightly different implementation than https://github.com/Sage-Bionetworks/genie-erbb2-cbio/blob/develop/create_meta.R#L220
}
\keyword{internal}
