% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{add_default_fileview}
\alias{add_default_fileview}
\title{Create default project fileview}
\usage{
add_default_fileview(project)
}
\arguments{
\item{project}{A project entity.}
}
\description{
Create default project fileview
}
