% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{add_default_wiki}
\alias{add_default_wiki}
\title{Add default wiki}
\usage{
add_default_wiki(
  project,
  name,
  pi,
  lead,
  funder,
  initiative,
  abstract,
  institution
)
}
\arguments{
\item{project}{Synapse id of project.}

\item{name}{Name of the project/study.}

\item{pi}{Name of the principal investigator.}

\item{lead}{Name(s) of the project lead/data coordinator, comma-sep if multiple, e.g. "Jane Doe, John Doe".}

\item{funder}{The funding agency. The relevant funder team will be made admin.}

\item{initiative}{Title of funding initiative, e.g. "Young Investigator Award".}

\item{abstract}{Project abstract/description.}

\item{institution}{Affiliated institution(s), \strong{semicolon-sep if multiple}, e.g. "Stanford University; University of California, San Francisco".}
}
\description{
Add the default wiki at project at creation or
use to retrofit projects where creators have not created a wiki.
}
