% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_people_table.R
\name{add_people_from_table}
\alias{add_people_from_table}
\title{Update the People table from a source Table or View column}
\usage{
add_people_from_table(
  people_table_id,
  people_column,
  source_table_id,
  source_column,
  dry_run = T
)
}
\arguments{
\item{people_table_id}{The synapse id of the table used for referencing people.}

\item{people_column}{Column name within the people table that contains the relevant people values.}

\item{source_table_id}{The synapse id of the source table.}

\item{source_column}{Column name within the source table that contains the relevant source values.}

\item{dry_run}{Default = TRUE Skips upload of annotations unless set to FALSE.}
}
\value{
If dry_run == T, prints preview of updated people table, otherwise uploads the updates.
}
\description{
Update the People table from a source Table or View column
}
