% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_study.R
\name{add_to_scope}
\alias{add_to_scope}
\title{Add to scope}
\usage{
add_to_scope(view_id, container_id)
}
\arguments{
\item{view_id}{Id of view}

\item{container_id}{Id of container to add.}
}
\description{
Convenience function to add container to view scope.
}
