% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_utils.R
\name{annotate_aligned_reads}
\alias{annotate_aligned_reads}
\title{Annotate processed aligned reads}
\usage{
annotate_aligned_reads(
  sample_io,
  samtools_stats_file = NULL,
  picard_stats_file = NULL,
  template = "bts:ProcessedAlignedReadsTemplate",
 
    schema = "https://raw.githubusercontent.com/nf-osi/nf-metadata-dictionary/main/NF.jsonld",
  verbose = TRUE,
  update = FALSE
)
}
\arguments{
\item{sample_io}{Table mapping input to outputs.}

\item{samtools_stats_file}{Path to file/syn id of file with samtools stats produced by the workflow.}

\item{picard_stats_file}{Path to file/syn id of file with picard stats produced by the workflow.}

\item{template}{(Optional) URI of template in data model to use, prefixed if needed.
Can specify different model/version, but in some cases may not work well.}

\item{schema}{Path (URL or local) to file from which schema will be read, or schema as list object.}

\item{verbose}{Give verbose reports for what's happening.}

\item{update}{Whether to apply annotations.}
}
\description{
Put together annotation components for nextflow star-salmon outputs. Annotations come from several sources:
\enumerate{
\item Inherit some annotations on the original input files. Requires a reference mapping of input files to use.
Most prop vals can be inherited by the derived files, e.g. assay type, but not for "comments" or "entityId".
Ideally, the data model itself should include inheritance rules; since that isn't possible currently,
we hard-code lots of stuff, so this is hard to generalize for other data models.
}
}
\details{
\enumerate{
\item Extract metrics from auxiliary files to surface as annotations. See  \code{\link{annotate_with_tool_stats}}.
\item Manually add annotations that can't (yet?) be derived from #1 or #2. Has to be done outside of this util.
}

Always returns a "partial" manifest; the param \code{update} specifies whether annotations should be applied.
}
