% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_utils.R
\name{annotate_called_variants}
\alias{annotate_called_variants}
\title{Annotate somatic or germline variants output}
\usage{
annotate_called_variants(
  sample_io,
  format = c("vcf", "maf"),
  template = "bts:ProcessedVariantCallsTemplate",
  schema =
    "https://raw.githubusercontent.com/nf-osi/nf-metadata-dictionary/main/NF.jsonld",
  data_type = c("SomaticVariants", "GermlineVariants"),
  verbose = TRUE,
  update = FALSE
)
}
\arguments{
\item{sample_io}{Table mapping input to outputs, where outputs are expected to be \code{.vcf.gz} files.}

\item{format}{Variant format, either "vcf" or "maf". See details for assumptions.}

\item{template}{(Optional) URI of template in data model to use, prefixed if needed.
Can specify different model/version, but in some cases may not work well.}

\item{schema}{Path (URL or local) to file from which schema will be read, or schema as list object.}

\item{data_type}{Variant type, given that this can be used with either somatic or germline.}

\item{verbose}{Give verbose reports for what's happening.}

\item{update}{Whether to apply annotations.}
}
\description{
Variant format defaults to \code{vcf}, with \code{maf}s produced in subsequent annotation workflow as is the SOP for NF.
Thus, if \code{maf}, this upgrades the dataType and overrides a few attribute values.
If in the future \code{maf}s require a different template, then this should be factored out into another specialized annotation function.
}
