% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrams.R
\name{as_mmd_node}
\alias{as_mmd_node}
\title{Generate notation for mermaid.js nodes}
\usage{
as_mmd_node(entity, class = c("Project", "Dataset", "Folder"))
}
\arguments{
\item{entity}{Character vector of one or more entity ids.
If named, nodes will use names instead of ids as labels. Note that entity ids
starting with "_" are considered blank nodes and are treated specially.}

\item{class}{Optional, add a class to the node.}
}
\description{
Generate notation for mermaid.js nodes
}
\keyword{internal}
