% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_study_data_types.R
\name{assign_study_data_types}
\alias{assign_study_data_types}
\title{Summarize data types for the study}
\usage{
assign_study_data_types(
  study_table_id,
  fileview_id,
  id_col = "studyId",
  attribute = "dataType",
  dry_run = TRUE
)
}
\arguments{
\item{study_table_id}{Synapse ID of reference portal study table. Used to get study ids.}

\item{fileview_id}{Synapse ID of the reference portal fileview.}

\item{id_col}{Name of the study id column in \code{study_table_id} and \code{fileview_id}. Defaults to \code{studyId}.}

\item{attribute}{Attribute being summarized using fileview. Defaults to \code{dataType}.}

\item{dry_run}{Default = TRUE. Whether to update as well or just return list of annotation objects.}
}
\value{
List of annotations objects.
}
\description{
Data types are summarized, or "rolled-up", for the study based on its child file annotations.
Summary values are added back as and overwrites the current \code{dataType} annotation for the study.
See also the related \code{update_study_annotations}, where study-level annotations are \emph{rolled down} to child files.
Note that under-the-hood this now wraps a generalized util \code{summarize_attribute}.
}
\examples{
\dontrun{
assign_study_data_types(study_table_id = 'syn52694652',
                        fileview_id = 'syn16858331',
                        id_col = 'studyId',
                        attribute = 'dataType',
                        dry_run = T)
}
}
