% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{bind_schema}
\alias{bind_schema}
\title{Wrapper for JSON schema binding}
\usage{
bind_schema(id, schema_id, derived_annotations = FALSE)
}
\arguments{
\item{id}{Id of entity to which schema will be bound}

\item{schema_id}{Schema id as registered on Synapse.}

\item{derived_annotations}{Whether to enabled derived annotations.
Default \code{FALSE} as this is the API default.}
}
\description{
See https://help.synapse.org/docs/JSON-Schemas.3107291536.html
}
