% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbioportal.R
\name{cbp_add_expression}
\alias{cbp_add_expression}
\title{Export and add expression data to cBioPortal dataset}
\usage{
cbp_add_expression(expression_data, expression_data_raw = NULL, verbose = TRUE)
}
\arguments{
\item{expression_data}{Syn id of normalized gene counts results (default to TPM). See details.}

\item{expression_data_raw}{(Optional) Syn id of raw counts results. See details.}

\item{verbose}{Whether to be chatty.}
}
\description{
This should be run in an existing dataset package root.
Note that there are a number of different options generated by the STAR Salmon pipeline.
cBioPortal has confirmed that they prefer normalized counts \code{gene_tpm.tsv} and,
\emph{though not used}, find it helpful to also have raw counts \code{gene_counts.tsv}.
}
