% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbioportal_study.R
\name{cbp_add_expression}
\alias{cbp_add_expression}
\title{Export and add expression data to cBioPortal dataset}
\usage{
cbp_add_expression(expression_data)
}
\arguments{
\item{expression_data}{Syn id of gene expression data.}
}
\description{
This should be run in an existing dataset package root.
}
