% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbioportal.R
\name{cbp_new_cancer_type}
\alias{cbp_new_cancer_type}
\title{Create index files for new cancer type}
\usage{
cbp_new_cancer_type(type_of_cancer, name, color, parent_type_of_cancer)
}
\description{
https://docs.cbioportal.org/file-formats/#cancer-type
}
