% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_testing_utils.R
\name{check_libtype_validity}
\alias{check_libtype_validity}
\title{Check output strandedness matches samplesheet strandedness}
\usage{
check_libtype_validity(
  syn_out,
  samplesheet,
  parse_fun = function(x) gsub("_T[0-9]$", "", x)
)
}
\arguments{
\item{samplesheet}{A local file or syn id of samplesheet.}

\item{parse_fun}{Function implementing how to parse samples in samplesheet.}
}
\description{
This function allows you to (post-workflow) check that the library type inferred by
Salmon (ISR, ISF, IU) during processing matches the "strandedness" provided for each sample in the workflow sample sheet.
}
\examples{
\dontrun{
 check_libtype_validity(syn_out = 'syn30840584', samplesheet = "syn30841441") #should pass
 check_libtype_validity(syn_out = 'syn30840584', samplesheet = "syn39593587") #should fail
}
}
