% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{cite_dataset}
\alias{cite_dataset}
\title{Generate example dataset citation}
\usage{
cite_dataset(id, format = "Scientific Data", output = c("markdown"))
}
\arguments{
\item{id}{Dataset id.}

\item{format}{Currently just "Scientific Data" format.}

\item{output}{Currently only markdown, from which other utils can be used to generate LaTeX or HTML.}
}
\description{
This is currently more for demo purposes, to check how well current metadata
could be formatted into citation text. Datasets where DOIs have been minted
\emph{or} NF-OSI processed datasets within the official Portal Collection should
work well, while there are no guarantees for other cases.
Note: Internal/experimental use only, not for production use.
}
\keyword{internal}
