% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_stringlist.R
\name{convert_to_stringlist}
\alias{convert_to_stringlist}
\title{Convert a delimited string to a stringlist annotation}
\usage{
convert_to_stringlist(
  fileview_id,
  annotation_key,
  sep = ",",
  trim_ws = TRUE,
  dry_run = TRUE
)
}
\arguments{
\item{fileview_id}{The synapse id of a fileview. Must have the desired annotations in the schema, and must have the files to annotate included in the scope. Must have write access to the files you want to re-annotate.}

\item{annotation_key}{A character string of the annotation you'd like to switch from a delimited string to a stringlist.}

\item{sep}{Default = ",". The delimiter in the character string.}

\item{dry_run}{Default = TRUE. Skips upload to table and instead prints study tibble.}

\item{trimws}{Default = TRUE. Remove white space at the beginning and end of list items (e.g. "NF1, NF2" and "NF1,NF2" will yield the same STRING_LIST result).}
}
\value{
If dry_run == T, returns study tibble and skips upload.
}
\description{
Converts a delimited string to a stringlist annotation and adjust the associated schema in the portal fileview.
}
