% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{copy_annotations}
\alias{copy_annotations}
\title{Copy annotations}
\usage{
copy_annotations(entity_from, entity_to, select = NULL, update = FALSE)
}
\arguments{
\item{entity_from}{Syn id from which to copy.}

\item{entity_to}{One or more syn ids to copy annotations to.}

\item{select}{Vector of properties to selectively copy if present on the entity.
If not specified, will copy over everything, which may not be desirable.}

\item{update}{Whether to immediately update or return annotation objects only.}
}
\description{
Copy annotations (all or selectively) from a source entity to one or more target entities.
If annotations already exist on target entities, the copy will replace the current values.
}
