% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_annotation_utils.R
\name{derive_annotations}
\alias{derive_annotations}
\title{Derive annotations for processed output data}
\usage{
derive_annotations(sample_io, template, schema, format, verbose = TRUE)
}
\arguments{
\item{format}{File format of the processed data, e.g. "vcf".}
}
\description{
A processed or derived file can inherit annotations from the input file(s).
Currently, this generously facilitates inheritance of many properties except ones that
"obviously" shouldn't be inherited, such as "fileFormat" or "comments".
These rules are hard-coded and might need to be expanded as the data model changes,
and the manifest generated should still be reviewed.
}
\details{
If multiple inputs given, this will inherit annotations from the FIRST input.
Files that pass through this naturally have \code{dataSubtype} automatically set to "processed" and \code{fileFormat} set
to the actual new file format. In the future, \code{template} itself may define default \code{format} so we don't need to specify explicitly.
}
\keyword{internal}
