% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_utils.R
\name{derive_annotations}
\alias{derive_annotations}
\title{Derive annotations for processed output data}
\usage{
derive_annotations(sample_io, template, schema, format, verbose = TRUE)
}
\arguments{
\item{format}{File format of the processed data.}
}
\description{
Build annotations through inheritance from inputs. If multiple inputs, inherit props from the FIRST input.
Files that pass through this naturally have \code{dataSubtype} set to "processed" and \code{fileFormat} set
to the actual new file format. In the future, \code{template} itself may define \code{format} so we don't need to specify explicitly.
}
\keyword{internal}
