% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_stringlist.R
\name{.replace_string_column_with_stringlist_column}
\alias{.replace_string_column_with_stringlist_column}
\title{Replace string column with stringlist column}
\usage{
.replace_string_column_with_stringlist_column(
  table_id,
  column_name,
  max_str_len
)
}
\arguments{
\item{table_id}{A synapse entity id.}

\item{column_name}{The column name of relevant column to modify.}

\item{max_str_len}{Max string length to be set in schema of new column.}
}
\description{
Guts of this ripped from @jaeddy gist (https://gist.github.com/jaeddy/1cf49f7851945beedb39d431134734af)
}
