% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_testing_utils.R
\name{extract_libtype}
\alias{extract_libtype}
\title{Extract the library type from cmd_info.json in nf-core/rnaseq output}
\usage{
extract_libtype(syn_id)
}
\arguments{
\item{syn_id}{A synapse id for cmd_info.json}
}
\value{
a string: ISR, ISF, or IU
}
\description{
Extract the library type from cmd_info.json in nf-core/rnaseq output
}
