% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_utils.R
\name{get_by_prop_from_json_schema}
\alias{get_by_prop_from_json_schema}
\title{Look up connected nodes by specified property in JSON-LD schema}
\usage{
get_by_prop_from_json_schema(
  id,
  prop,
 
    schema = "https://raw.githubusercontent.com/nf-osi/nf-metadata-dictionary/main/NF.jsonld",
  return_labels = TRUE,
  recursive = FALSE,
  result = NULL,
  rest = NULL
)
}
\arguments{
\item{id}{Id (\verb{@id}) for which to get range values; include prefix if needed.}

\item{prop}{Property; include prefix if needed.}

\item{schema}{Path (URL or local) to file from which schema will be read, or schema as list object.}

\item{return_labels}{Return labels (default), otherwise ids of connected nodes.}

\item{recursive}{Recursive lookup?}

\item{result}{Vector of accumulated results; used for recursive lookup.}

\item{rest}{Vector of remaining ids; used for recursive lookup.}
}
\description{
Given \verb{@id}, looks up connected nodes by specified property (e.g. 'sms:something').
Intended to be a generic used to define more specific lookup conveniences.
Can do recursive lookup (which makes sense for certain properties such as "dependsOn"
but doesn't make sense for properties such as "rdfs:label");
this assumes that the graph is actually a tree (acyclic).
Also the JSON-LD schema is assumed to be schematic-generated,
and this might not work with JSON-LD generated from other tools such as LinkML
due to small differences and hard-coded stuff.
}
