% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{get_cbio_filename}
\alias{get_cbio_filename}
\title{Get cBioPortal clinical file name based on clinical data type}
\usage{
get_cbio_filename(clinical_type = c("SAMPLE", "PATIENT"))
}
\arguments{
\item{clinical_type}{String representing cBioPortal clinical data type.}
}
\value{
string
}
\description{
This is called from the wrapper \code{write_cbio_clinical}.
Adapted from https://github.com/Sage-Bionetworks/genie-erbb2-cbio/blob/develop/create_clinical.R#L411.
Note that of the clinical file types, the PATIENT type can actually be optional,
and we (NF) currently don't use the TIMELINE type, so options have been simplified.
}
