% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_study_data_types.R
\name{get_valid_values_from_json_schema}
\alias{get_valid_values_from_json_schema}
\title{Retrieve valid subclasses of a value in a JSON-LD schema}
\usage{
get_valid_values_from_json_schema(
 
    schema_url = "https://raw.githubusercontent.com/nf-osi/schematic/develop/data/schema_org_schemas/NF.jsonld",
  parent_name = "DataType",
  parent_context = "bts"
)
}
\arguments{
\item{schema_url}{Default: the NF-OSI JSON-LD schema.}

\item{parent_name}{Default = DataType. The value for which you'd like to find the associated subclasses.}

\item{parent_context}{Default = bts. The JSON-LD context for the value in question.}
}
\value{
A character vector of values.
}
\description{
Retrieve valid subclasses of a value in a JSON-LD schema generated by schematic.
}
