% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_qc.R
\name{infer_data_type}
\alias{infer_data_type}
\title{Infer data type of a dataset folder}
\usage{
infer_data_type(dataset_id)
}
\arguments{
\item{dataset_id}{Optional, if given this fills out manifest for existing dataset instead of generating a blank manifest.}
}
\value{
List of structure \code{list(result = result, notes = notes)}, where \code{result} can be \code{NA}.
}
\description{
Infer the data type by checking the first few files.
TODO: Check \code{dataType} instead of Component and derive Component
because some older files does not have Component explicitly.
}
