% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{is_dataset_collection}
\alias{is_dataset_collection}
\title{Check whether entity is dataset collection}
\usage{
is_dataset_collection(id)
}
\description{
Check whether entity is dataset collection
}
\keyword{internal}
