% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{make_admin}
\alias{make_admin}
\title{Make a user or group full admin of a Synapse entity}
\usage{
make_admin(entity, principal_id)
}
\arguments{
\item{entity}{The Synapse entity, e.g. project or folder.}

\item{principal_id}{User/team name or id (e.g. "NF-OSI Sage Team", "3378999", "nf-bot", or "3423450") that will have the configured access to the entity.}
}
\description{
Convenience method to set admin permissions
}
