% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{make_cbio_clinical_header}
\alias{make_cbio_clinical_header}
\title{Make header for cBioPortal clinical data file}
\usage{
make_cbio_clinical_header(df, label, description, data_type)
}
\arguments{
\item{df}{A \code{data.frame} representing clinical dataset to publicize.}

\item{label}{Character vector representing a short label for each column in the dataset}

\item{description}{Character vector representing a long descriptions for each column in the dataset}

\item{data_type}{Character vector representing the data type of each column in the dataset}
}
\description{
This is called from the wrapper \code{write_cbio_clinical}.
Reused from https://github.com/Sage-Bionetworks/genie-erbb2-cbio/blob/develop/create_clinical.R#L396.
}
