% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{make_folder}
\alias{make_folder}
\title{Create project folders}
\usage{
make_folder(parent, folders)
}
\arguments{
\item{parent}{The Synapse id or object that should be the parent container, i.e. the project or another folder.}

\item{folders}{List giving one or more folder names of folder(s) to create.}
}
\value{
A list of the created folder object(s).
}
\description{
Use to set up a scaffold of standard upper-level folders as well as
customized data folders within "Raw Data" for a new project.
}
\examples{
\dontrun{
datasets <- list("sequencing data", "imaging data")
assays <- c("rnaSeq", "immunohistochemistry")
for(i in seq_along(datasets)) attr(datasets[[i]], "assay") <- assays[[i]]
make_folder(parent = "syn26462036", datasets)
}
}
