% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{make_meta_study}
\alias{make_meta_study}
\title{Make meta study file}
\usage{
make_meta_study(
  cancer_study_identifier,
  type_of_cancer = "mixed",
  name,
  description,
  citation = NULL,
  pmid = NULL,
  groups = "PUBLIC",
  short_name = NULL,
  add_global_case_list = TRUE,
  publish_dir = ".",
  write = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{cancer_study_identifier}{The study identifier.}

\item{type_of_cancer}{Type of cancer, defaults to "mixed". See also http://oncotree.mskcc.org/#/home.}

\item{name}{Name of the study.}

\item{description}{Description of the study.}

\item{citation}{(Optional) A relevant citation, e.g. "TCGA, Nature 2012".}

\item{pmid}{(Optional) One or more relevant pubmed ids (comma separated without whitespace); if used, citation cannot be \code{NULL}.}

\item{groups}{(Optional) Defaults to "PUBLIC" for use with public cBioPortal;
otherwise, use group names that make sense for your instance.}

\item{short_name}{(Optional) Short name for the study.}

\item{add_global_case_list}{(Optional) Use \code{NULL} to ignore, but default is \code{TRUE} for an "All samples" case list to be generated automatically.}

\item{publish_dir}{Directory path to write to, defaults to current.}

\item{write}{Whether to write the meta file for the clinical data file.}

\item{verbose}{Report where file has been written.}
}
\description{
Adapted from https://github.com/Sage-Bionetworks/genie-erbb2-cbio/blob/develop/create_meta.R#L179
See specifications at https://docs.cbioportal.org/file-formats/#meta-file.
}
