% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{make_meta_study_generic}
\alias{make_meta_study_generic}
\title{Template for meta study file}
\usage{
make_meta_study_generic(
  type_of_cancer,
  cancer_study_identifier,
  name,
  description,
  citation = NULL,
  pmid = NULL,
  groups = NULL,
  short_name = NULL,
  add_global_case_list = NULL
)
}
\description{
Adapted from https://github.com/Sage-Bionetworks/genie-erbb2-cbio/blob/develop/create_meta.R#L90
Low-level internal function for the tedious templating.
}
\keyword{internal}
