% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_qc.R
\name{manifest_validate}
\alias{manifest_validate}
\title{Validate manifest via schematic service}
\usage{
manifest_validate(
  data_type,
  json_str = NULL,
  restrict_rules = FALSE,
  schema_url =
    "https://raw.githubusercontent.com/nf-osi/nf-metadata-dictionary/main/NF.jsonld",
  service = "https://schematic.api.sagebionetworks.org/v1/model/validate"
)
}
\arguments{
\item{data_type}{Data type of the manifest to generate (aka Component).}

\item{json_str}{JSON string representing metadata.}

\item{restrict_rules}{Use only basic schematic validation instead of extended validation with Great Expectations, default \code{FALSE}.}

\item{schema_url}{Optional, defaults to main NF 'latest' data model.}

\item{service}{Service endpoint to use. Defaults to the schematic production endpoint.}
}
\description{
Get validation results from schematic service. Some downstream utils consume these results for custom display or summary.
NOTE: This doesn't support files, though the endpoint supports uploading \code{.csv}.
}
