% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_annotation_utils.R
\name{map_sample_input_ss}
\alias{map_sample_input_ss}
\title{Parse nextflow samplesheet for sample inputs}
\usage{
map_sample_input_ss(
  samplesheet,
  parse_fun = function(x) gsub("_T[0-9]$", "", x)
)
}
\arguments{
\item{samplesheet}{A local file or syn id of samplesheet.}

\item{parse_fun}{Function implementing how to parse samples in samplesheet.}
}
\description{
Samplesheets are used in rnaseq pipelines, defined here: https://nf-co.re/rnaseq/usage#full-samplesheet.
After the pipeline is run, it should be found in an output folder called \code{pipeline_info}.
This is a simple helper to get a mapping of sample ids to input files (either one-to-many or one-to-one) as a table.
}
