% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_annotation_utils.R
\name{map_sample_output_rnaseq}
\alias{map_sample_output_rnaseq}
\title{Map sample to output from nf-rnaseq}
\usage{
map_sample_output_rnaseq(syn_out)
}
\arguments{
\item{syn_out}{Syn id of syn output destination with files of interest.}
}
\value{
A \code{data.table} with cols \code{output_name} \code{output_id} \code{sample} \code{workflow}
}
\description{
See https://nf-co.re/rnaseq. Given location where workflow has deposited the outputs,
map relevant processed files based on extensions and link these files to source samples.
So for \verb{results/star_salmon/<sample>/<file>}, the URI to pass in is the \code{star_salmon} folder.
Warning: Reliance on certain file structure and naming convention can make this somewhat brittle!
}
\details{
See the related \code{\link{map_sample_input_ss}} for mapping sample to inputs instead of outputs.
}
