% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_utils.R
\name{map_sample_output_sarek}
\alias{map_sample_output_sarek}
\title{Map sample to output from nf-sarek}
\usage{
map_sample_output_sarek(syn_out)
}
\arguments{
\item{syn_out}{Syn id of syn output destination with files of interest.}
}
\value{
A \code{data.table} with cols \code{caller} \code{caller_path} \code{caller_syn} \code{output_name} \code{output_id} \code{sample} \code{workflow}
}
\description{
See https://nf-co.re/sarek. Processed outputs are nested by sample and variant callers, i.e.
\verb{*VariantCalling/<TUMOR_vs_NORMAL>/<CALLER>}. This walks through the output destination (URI of \verb{*VariantCalling})
with similar intention to \code{\link{map_sample_output_rnaseq}}, but for Sarek outputs.
}
