% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_utils.R
\name{map_sample_output_sarek}
\alias{map_sample_output_sarek}
\title{Map sample to output from nf-sarek}
\usage{
map_sample_output_sarek(syn_out)
}
\arguments{
\item{syn_out}{Syn id of syn output destination (folder) with files of interest.}
}
\value{
A \code{data.table} with cols \code{caller} \code{caller_path} \code{caller_syn} \code{output_name}
\code{output_name} \code{output_id} \code{sample}
}
\description{
See https://nf-co.re/sarek.
Processed outputs are nested by sample and different variant callers
with structure \verb{VariantCalling/[TUMOR_vs_NORMAL]/[CALLER]}.
This looks through the \verb{*VariantCalls} output directory and attempts to map
the relevant processed files based on file extensions and tie these files to source samples.
This relies on expected file locations and file naming convention of the specific workflow.
WARNING: Be aware that this may not work using a different version of workflow
or if files are manually re-organized post-write to Synapse.
}
