% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_annotation_utils.R
\name{map_sample_output_sarek}
\alias{map_sample_output_sarek}
\title{Map sample to output from nf-sarek}
\usage{
map_sample_output_sarek(syn_out, sample_level = 2)
}
\arguments{
\item{syn_out}{Syn id of syn output destination with files of interest.}

\item{sample_level}{If caller is organized by sample, use 2 (default), if samples organized by caller, use 3. See details.}
}
\value{
A \code{data.table} with cols \code{caller} \code{caller_path} \code{caller_syn} \code{output_name} \code{output_id} \code{sample} \code{workflow}
}
\description{
See https://nf-co.re/sarek. Most processed outputs are nested by sample and variant callers, i.e.
\verb{*VariantCalling/<TUMOR_vs_NORMAL>/<CALLER>}. Other times the data is organized as
\verb{*VariantCalling/<CALLER>/<TUMOR_vs_NORMAL>}.
This walks through the output destination (URI of \verb{*VariantCalling})
with similar intention to \code{\link{map_sample_output_rnaseq}}, but for Sarek outputs.
}
\details{
Note: And additional step post-Sarek will create MAFs in the output subdirectory DeepVariant.
If this is run \emph{after} the MAF creation step, this will return file indexes with those .maf files.
}
