% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_qc.R
\name{meta_qc_project}
\alias{meta_qc_project}
\title{QC metadata at the project level with pass/fail result}
\usage{
meta_qc_project(project_id)
}
\arguments{
\item{project_id}{Synapse project id.}
}
\value{
A table of with rows for the datasets QC'd, with dataset id, name, TRUE/FALSE pass result, and summary;
otherwise \code{NA}.
}
\description{
For projects with a relatively standard structure that also corresponds to what the DCA expects,
this is an adequate wrapper to go through the datasets and do basic QC in a one-stop-shop manner.
For selective validation or other (e.g. milestone-based or ad hoc) structures, look at \code{meta_qc_dataset}.
}
