% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_qc.R
\name{meta_qc_project}
\alias{meta_qc_project}
\title{QC metadata at the project level with pass/fail result}
\usage{
meta_qc_project(project_id, result_file = NULL, ...)
}
\arguments{
\item{project_id}{Synapse project id.}

\item{result_file}{If not NULL, \emph{also} write to output to \code{.csv} file.}

\item{...}{Params passed to \code{meta_qc_dataset}.}
}
\value{
A table of with rows for the datasets QC'd, with dataset id, name, TRUE/FALSE pass result, and summary;
otherwise \code{NA}.
}
\description{
An adequate wrapper to go through project datasets and do basic QC in one-stop-shop manner
\strong{for projects that have standard structure corresponding to what DCA expects}.
}
\details{
For selective validation or other (e.g. milestone-based) structures, look at \code{meta_qc_dataset}.
}
