% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_annotation_utils.R
\name{processed_meta}
\alias{processed_meta}
\title{Metadata for processed products}
\usage{
processed_meta(input, output, workflow, workflow_link)
}
\arguments{
\item{input}{Data from \code{map_sample_input_ss}.}

\item{output}{Data from \verb{map_sample_output_*}.}

\item{workflow}{Either "nf-rnaseq" or "nf-sarek".}

\item{workflow_link}{Workflow link.}
}
\value{
List \code{manifest} with manifests for each processed dataset,
and \code{sample_io} with linked inputs and outputs (should be used for provenance annotation).
}
\description{
Wrapper for all steps to get manifest for processed product
}
