% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrams.R
\name{processing_flowchart}
\alias{processing_flowchart}
\title{Wrapper to create data-driven flowchart with pretty processing provenance mermaid template}
\usage{
processing_flowchart(report)
}
\arguments{
\item{report}{Which report to determine subset of data for which to generate flowchart fig.}
}
\description{
This generates a mermaid file, which can be rendered live/further edited in
}
\examples{
\dontrun{
flowchart <- processing_flowchart(report = "2023-MY")
cat(flowchart, file = "flowchart.mmd")
}
}
