% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_utils.R
\name{qc_manifest}
\alias{qc_manifest}
\title{QC derived manifest}
\usage{
qc_manifest(manifest, sample_io, default = "sex")
}
\arguments{
\item{manifest}{The manifest from one of the \verb{annotate_*} functions.}

\item{sample_io}{The input/output mapping used.}

\item{default}{Default attribute to use as canary to check quality/view problematic input/outputs.}
}
\description{
Check missing annotations, usually because input files are missing values so nothing could be copied.
}
