% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{use_latest_in_collection}
\alias{use_latest_in_collection}
\title{Update item versions to "latest" in a collection}
\usage{
use_latest_in_collection(
  collection_id,
  items = "all",
  version_semantics = "abs"
)
}
\arguments{
\item{collection_id}{Collection id.}

\item{items}{Vector of dataset ids for which to update reference to latest version, or "all" (default) to update all.}

\item{version_semantics}{Use "abs" for absolute latest version or "stable". Only used for collection entities. See details.}
}
\description{
Update an \emph{existing} collection so that all items or a subset of items reference their latest version.
Should work for both datasets (collection of files) and dataset collections (collection of datasets).
}
