% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{use_latest_in_collection}
\alias{use_latest_in_collection}
\title{Update item versions to "latest" in a collection}
\usage{
use_latest_in_collection(collection_id, items = "all")
}
\arguments{
\item{collection_id}{Collection id.}

\item{items}{Vector of dataset ids for which to update reference to latest version,
or "all" (default) to update all in the dataset collection.}
}
\description{
Update an \emph{existing} collection so that all items or a subset of items reference their latest version.
This should work for both datasets (collection of files) and dataset collections (collection of datasets).
}
