% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_collection_items}
\alias{validate_collection_items}
\title{Validate metadata of items in a collection}
\usage{
validate_collection_items(collection_id, schema_id)
}
\arguments{
\item{collection_id}{Collection id.}

\item{schema_id}{Id of schema that items are expected to be bound to, e.g. "org.synapse.nf-portaldataset".}
}
\description{
This is usually used with a dataset collection.
For each item in the collection, check first that it is bound to the expected schema.
Then check that the item's annotations validate against the schema, using Synapse's native validation services.
}
\details{
The result partitions any \code{bound_schema_error} and \code{validation_error} in a list.
}
