% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboilerplate.R
\name{write_cbio_clinical}
\alias{write_cbio_clinical}
\title{Write cBioPortal clinical file}
\usage{
write_cbio_clinical(
  df,
  ref_map,
  na_recode = c("NA", "NaN", "unknown", "Unknown"),
  delim = "\\t",
  publish_dir = ".",
  verbose = TRUE
)
}
\arguments{
\item{df}{A \code{data.frame} representing clinical dataset to publicize.}

\item{ref_map}{YAML or JSON mapping. See details.}

\item{na_recode}{Possible NA values to replace with a blank string (which seems to be standard) in exported file.}

\item{delim}{Delimiter character used for writing file, defaults to tab-delimited per cBioPortal specs.}

\item{publish_dir}{Directory path to write to, defaults to current.}

\item{verbose}{Report where file has been written.}
}
\description{
Wrapper function for creating clinical files. There are two: PATIENT and SAMPLE.
The PATIENT file is actually optional, so there are only checks for making sure SAMPLE can be created.
\code{df} is expected to be a table containing clinical data available, and maybe even some irrelevant data
(since NF data is not well-normalized and there is a single table with everything).
}
\details{
This depends on a \code{ref_map} specification to know which clinical data to include for cBioPortal
and how to segregate the clinical attributes into the right files.
Basically, \code{ref_map} decides what variables can be made public and how they should be represented in cBioPortal.
For example, given a table \code{T} on Synapse with variables A-Z and mappings in \code{ref_map} for A-C + L-M,
we take the intersection of variables present.
But first, check that \emph{required} variables in \emph{ref_map} are present.
So first the subset \code{df} is created from \code{T}.
}
\keyword{internal}
